+++ src/main/java/com/btjf/controller/productpm/ProductPmController.java
@@ -5,7 +5,7 @@ import com.btjf.application.components.xaresult.AppXaResultHelper;
 import com.btjf.application.util.XaResult;
 import com.btjf.common.page.Page;
 import com.btjf.controller.base.ProductBaseController;
-import com.btjf.model.product.Productpm;
+import com.btjf.model.product.ProductPm;
 import com.btjf.service.pm.PmService;
 import com.btjf.service.productpm.ProductPmService;
 import com.wordnik.swagger.annotations.Api;
@@ -25,7 +25,7 @@ import java.util.List;
 @RestController
 @RequestMapping(value = "/productpm")
 @Api(value = "ProductPmController", description = "耗料管理", position = 1)
-public class ProductPmController extends ProductBaseController{
+public class ProductPmController extends ProductBaseController {
 
     private static final Logger LOGGER = Logger
             .getLogger(ProductPmController.class);
@@ -37,50 +37,50 @@ public class ProductPmController extends ProductBaseController{
     private PmService pmService;
 
     @RequestMapping(value = "/list", method = RequestMethod.GET)
-    public XaResult<List<Productpm>> findList(@ApiParam("编号") String productNo, @ApiParam("名称") String pmNo
+    public XaResult<List<ProductPm>> findList(@ApiParam("编号") String productNo, @ApiParam("名称") String pmNo
             , @ApiParam("1已确认  0 未确认") int status, Integer pageSize, Integer currentPage) {
         getLoginUser();
         LOGGER.info(getRequestParamsAndUrl());
 
-        Page<Productpm> listPage = productPmService.findListPage(productNo, pmNo, status, AppPageHelper.appInit(currentPage, pageSize));
-        XaResult<List<Productpm>> result = AppXaResultHelper.success(listPage, listPage.getRows());
+        Page<ProductPm> listPage = productPmService.findListPage(productNo, pmNo, status, AppPageHelper.appInit(currentPage, pageSize));
+        XaResult<List<ProductPm>> result = AppXaResultHelper.success(listPage, listPage.getRows());
         return result;
     }
 
     @RequestMapping(value = "/modelConfige", method = RequestMethod.POST)
-    public XaResult<Integer> modelConfige(Integer[] ids){
+    public XaResult<Integer> modelConfige(Integer[] ids) {
         getLoginUser();
         LOGGER.info(getRequestParamsAndUrl());
 
         if (null == ids || Arrays.asList(ids).size() <= 0) {
             return XaResult.error("请选择要确认的记录");
         }
-        Integer row =  productPmService.updateStatue(Arrays.asList(ids));
+        Integer row = productPmService.updateStatue(Arrays.asList(ids));
         return XaResult.success(row);
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
-    public XaResult<Integer> delete(Integer[] ids){
+    public XaResult<Integer> delete(Integer[] ids) {
         getLoginUser();
         LOGGER.info(getRequestParamsAndUrl());
 
         if (null == ids || Arrays.asList(ids).size() <= 0) {
             return XaResult.error("请选择要删除的记录");
         }
-        Integer row =  productPmService.delete(Arrays.asList(ids));
+        Integer row = productPmService.delete(Arrays.asList(ids));
         return XaResult.success(row);
     }
 
     @RequestMapping(value = "/detail", method = RequestMethod.POST)
-    public XaResult<Productpm> detail(Integer id){
+    public XaResult<ProductPm> detail(Integer id) {
         getLoginUser();
         LOGGER.info(getRequestParamsAndUrl());
 
-       if(id == null){
-           return XaResult.error("请选择要查看的记录");
-       }
-        Productpm productpm = productPmService.getByID(id);
-        if(null == productpm){
+        if (id == null) {
+            return XaResult.error("请选择要查看的记录");
+        }
+        ProductPm productpm = productPmService.getByID(id);
+        if (null == productpm) {
             return XaResult.error("查看的记录不存在");
         }
         return XaResult.success(productpm);
@@ -88,25 +88,28 @@ public class ProductPmController extends ProductBaseController{
 
 
     @RequestMapping(value = "/addOrUpdate", method = RequestMethod.POST)
-    public XaResult<Productpm> updateOrAdd(Integer id, String productNo, String pmNo, String pmName , String num ,
-                                           String unit, String unitNum, String type, String remark, int status){
-      /*  getLoginUser();
+    public XaResult<ProductPm> updateOrAdd(Integer id, String productNo, String pmNo, String pmName, String num,
+                                           String unit, String unitNum, String type, String remark, int status) {
+        getLoginUser();
         LOGGER.info(getRequestParamsAndUrl());
 
-        if(null == pmService.getByNo(pmNo)) {
+        if (null == pmService.getByNo(pmNo)) {
             return XaResult.error("物料编号填写错误，请修改");
         }
 
-       // if()
 
-        if(status != 0 || status != 1){ return XaResult.error("是否确认类型错误");}
+        if (status != 0 || status != 1) {
+            return XaResult.error("是否确认类型错误");
+        }
 
-        if(id != null){
+        if (id != null) {
 
-        }else{
-
-        }*/
-      return null;
+        } else {
+            if (null != productPmService.getByNo(productNo)) {
+                return XaResult.error("物料编号已经存在");
+            }
+        }
+        return null;
     }
