+++ src/main/java/com/btjf/mapper/customer/CustomerMapper.xml
@@ -157,18 +157,21 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_customer (id, name, industry, 
-      salesman, source, level, 
-      product, linkMan, phone, 
-      tel, fax, email, address, 
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_customer (name, industry, salesman, 
+      source, level, product, 
+      linkMan, phone, tel, 
+      fax, email, address, 
       qq, msn, bankName, 
       bankNo, legalPerson, duty, 
       companyAddress, remark, createTime, 
       lastModifyTime, isDelete)
-    values (#{id,jdbcType=INTEGER}, #{name,jdbcType=VARCHAR}, #{industry,jdbcType=VARCHAR}, 
-      #{salesman,jdbcType=VARCHAR}, #{source,jdbcType=VARCHAR}, #{level,jdbcType=VARCHAR}, 
-      #{product,jdbcType=VARCHAR}, #{linkMan,jdbcType=VARCHAR}, #{phone,jdbcType=VARCHAR}, 
-      #{tel,jdbcType=VARCHAR}, #{fax,jdbcType=VARCHAR}, #{email,jdbcType=VARCHAR}, #{address,jdbcType=VARCHAR}, 
+    values (#{name,jdbcType=VARCHAR}, #{industry,jdbcType=VARCHAR}, #{salesman,jdbcType=VARCHAR}, 
+      #{source,jdbcType=VARCHAR}, #{level,jdbcType=VARCHAR}, #{product,jdbcType=VARCHAR}, 
+      #{linkMan,jdbcType=VARCHAR}, #{phone,jdbcType=VARCHAR}, #{tel,jdbcType=VARCHAR}, 
+      #{fax,jdbcType=VARCHAR}, #{email,jdbcType=VARCHAR}, #{address,jdbcType=VARCHAR}, 
       #{qq,jdbcType=VARCHAR}, #{msn,jdbcType=VARCHAR}, #{bankName,jdbcType=VARCHAR}, 
       #{bankNo,jdbcType=VARCHAR}, #{legalPerson,jdbcType=VARCHAR}, #{duty,jdbcType=VARCHAR}, 
       #{companyAddress,jdbcType=VARCHAR}, #{remark,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, 
@@ -179,11 +182,11 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_customer
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="name != null">
         name,
       </if>
@@ -255,9 +258,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="name != null">
         #{name,jdbcType=VARCHAR},
       </if>
