+++ src/main/java/com/btjf/mapper/dictionary/DictionaryMapper.xml
@@ -134,21 +134,22 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_dictionary (id, type, value
-      )
-    values (#{id,jdbcType=INTEGER}, #{type,jdbcType=INTEGER}, #{value,jdbcType=VARCHAR}
-      )
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_dictionary (type, value)
+    values (#{type,jdbcType=INTEGER}, #{value,jdbcType=VARCHAR})
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.dictionary.Dictionary">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_dictionary
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="type != null">
         type,
       </if>
@@ -157,9 +158,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="type != null">
         #{type,jdbcType=INTEGER},
       </if>
@@ -239,7 +237,7 @@
     where id = #{id,jdbcType=INTEGER}
   </update>
 
-  <select id = "getList" resultType="java.lang.String">
+  <select id="getList" resultType="java.lang.String">
     select value from t_dictionary where type = #{type,jdbcType=INTEGER}
   </select>
 </mapper>
