+++ src/main/java/com/btjf/mapper/emp/EmpMapper.xml
@@ -13,7 +13,7 @@
     <result column="sex" jdbcType="INTEGER" property="sex" />
     <result column="workId" jdbcType="INTEGER" property="workId" />
     <result column="type" jdbcType="INTEGER" property="type" />
-    <result column="native" jdbcType="VARCHAR" property="native" />
+    <result column="nativeSource" jdbcType="VARCHAR" property="nativeSource" />
     <result column="Nation" jdbcType="VARCHAR" property="nation" />
     <result column="birthday" jdbcType="VARCHAR" property="birthday" />
     <result column="idCard" jdbcType="VARCHAR" property="idCard" />
@@ -110,7 +110,7 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    id, name, deptId, isLeader, sex, workId, type, native, Nation, birthday, idCard, 
+    id, name, deptId, isLeader, sex, workId, type, nativeSource, Nation, birthday, idCard, 
     entryDate, school, subject, education, linkMan, phone, tel, address, salary, ylbx, 
     sybx, yiliaobx, gjj, dhbt, phoneSubsidy, foodSubsidy, socialSubsidy, remark, createTime, 
     lastModifyTime, isDelete
@@ -166,39 +166,42 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_emp (id, name, deptId, 
-      isLeader, sex, workId, 
-      type, native, Nation, 
-      birthday, idCard, entryDate, 
-      school, subject, education, 
-      linkMan, phone, tel, 
-      address, salary, ylbx, 
-      sybx, yiliaobx, gjj, 
-      dhbt, phoneSubsidy, foodSubsidy, 
-      socialSubsidy, remark, createTime, 
-      lastModifyTime, isDelete)
-    values (#{id,jdbcType=INTEGER}, #{name,jdbcType=VARCHAR}, #{deptId,jdbcType=INTEGER}, 
-      #{isLeader,jdbcType=INTEGER}, #{sex,jdbcType=INTEGER}, #{workId,jdbcType=INTEGER}, 
-      #{type,jdbcType=INTEGER}, #{native,jdbcType=VARCHAR}, #{nation,jdbcType=VARCHAR}, 
-      #{birthday,jdbcType=VARCHAR}, #{idCard,jdbcType=VARCHAR}, #{entryDate,jdbcType=VARCHAR}, 
-      #{school,jdbcType=VARCHAR}, #{subject,jdbcType=VARCHAR}, #{education,jdbcType=VARCHAR}, 
-      #{linkMan,jdbcType=VARCHAR}, #{phone,jdbcType=VARCHAR}, #{tel,jdbcType=VARCHAR}, 
-      #{address,jdbcType=VARCHAR}, #{salary,jdbcType=DECIMAL}, #{ylbx,jdbcType=DECIMAL}, 
-      #{sybx,jdbcType=DECIMAL}, #{yiliaobx,jdbcType=DECIMAL}, #{gjj,jdbcType=DECIMAL}, 
-      #{dhbt,jdbcType=DECIMAL}, #{phoneSubsidy,jdbcType=DECIMAL}, #{foodSubsidy,jdbcType=DECIMAL}, 
-      #{socialSubsidy,jdbcType=DECIMAL}, #{remark,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, 
-      #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_emp (name, deptId, isLeader, 
+      sex, workId, type, 
+      nativeSource, Nation, birthday, 
+      idCard, entryDate, school, 
+      subject, education, linkMan, 
+      phone, tel, address, 
+      salary, ylbx, sybx, 
+      yiliaobx, gjj, dhbt, 
+      phoneSubsidy, foodSubsidy, socialSubsidy, 
+      remark, createTime, lastModifyTime, 
+      isDelete)
+    values (#{name,jdbcType=VARCHAR}, #{deptId,jdbcType=INTEGER}, #{isLeader,jdbcType=INTEGER}, 
+      #{sex,jdbcType=INTEGER}, #{workId,jdbcType=INTEGER}, #{type,jdbcType=INTEGER}, 
+      #{nativeSource,jdbcType=VARCHAR}, #{nation,jdbcType=VARCHAR}, #{birthday,jdbcType=VARCHAR}, 
+      #{idCard,jdbcType=VARCHAR}, #{entryDate,jdbcType=VARCHAR}, #{school,jdbcType=VARCHAR}, 
+      #{subject,jdbcType=VARCHAR}, #{education,jdbcType=VARCHAR}, #{linkMan,jdbcType=VARCHAR}, 
+      #{phone,jdbcType=VARCHAR}, #{tel,jdbcType=VARCHAR}, #{address,jdbcType=VARCHAR}, 
+      #{salary,jdbcType=DECIMAL}, #{ylbx,jdbcType=DECIMAL}, #{sybx,jdbcType=DECIMAL}, 
+      #{yiliaobx,jdbcType=DECIMAL}, #{gjj,jdbcType=DECIMAL}, #{dhbt,jdbcType=DECIMAL}, 
+      #{phoneSubsidy,jdbcType=DECIMAL}, #{foodSubsidy,jdbcType=DECIMAL}, #{socialSubsidy,jdbcType=DECIMAL}, 
+      #{remark,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
+      #{isDelete,jdbcType=INTEGER})
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.emp.Emp">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_emp
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="name != null">
         name,
       </if>
@@ -217,8 +220,8 @@
       <if test="type != null">
         type,
       </if>
-      <if test="native != null">
-        native,
+      <if test="nativeSource != null">
+        nativeSource,
       </if>
       <if test="nation != null">
         Nation,
@@ -294,9 +297,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="name != null">
         #{name,jdbcType=VARCHAR},
       </if>
@@ -315,8 +315,8 @@
       <if test="type != null">
         #{type,jdbcType=INTEGER},
       </if>
-      <if test="native != null">
-        #{native,jdbcType=VARCHAR},
+      <if test="nativeSource != null">
+        #{nativeSource,jdbcType=VARCHAR},
       </if>
       <if test="nation != null">
         #{nation,jdbcType=VARCHAR},
@@ -430,8 +430,8 @@
       <if test="record.type != null">
         type = #{record.type,jdbcType=INTEGER},
       </if>
-      <if test="record.native != null">
-        native = #{record.native,jdbcType=VARCHAR},
+      <if test="record.nativeSource != null">
+        nativeSource = #{record.nativeSource,jdbcType=VARCHAR},
       </if>
       <if test="record.nation != null">
         Nation = #{record.nation,jdbcType=VARCHAR},
@@ -523,7 +523,7 @@
       sex = #{record.sex,jdbcType=INTEGER},
       workId = #{record.workId,jdbcType=INTEGER},
       type = #{record.type,jdbcType=INTEGER},
-      native = #{record.native,jdbcType=VARCHAR},
+      nativeSource = #{record.nativeSource,jdbcType=VARCHAR},
       Nation = #{record.nation,jdbcType=VARCHAR},
       birthday = #{record.birthday,jdbcType=VARCHAR},
       idCard = #{record.idCard,jdbcType=VARCHAR},
@@ -577,8 +577,8 @@
       <if test="type != null">
         type = #{type,jdbcType=INTEGER},
       </if>
-      <if test="native != null">
-        native = #{native,jdbcType=VARCHAR},
+      <if test="nativeSource != null">
+        nativeSource = #{nativeSource,jdbcType=VARCHAR},
       </if>
       <if test="nation != null">
         Nation = #{nation,jdbcType=VARCHAR},
@@ -667,7 +667,7 @@
       sex = #{sex,jdbcType=INTEGER},
       workId = #{workId,jdbcType=INTEGER},
       type = #{type,jdbcType=INTEGER},
-      native = #{native,jdbcType=VARCHAR},
+      nativeSource = #{nativeSource,jdbcType=VARCHAR},
       Nation = #{nation,jdbcType=VARCHAR},
       birthday = #{birthday,jdbcType=VARCHAR},
       idCard = #{idCard,jdbcType=VARCHAR},
