+++ src/main/java/com/btjf/mapper/emp/EmpSalaryMonthlyMapper.xml
@@ -152,31 +152,34 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_emp_salary_monthly (id, yearMonth, empId, 
-      empName, deptId, deptName, 
-      workName, salary, workDay, 
-      dayWork, nightWork, score, 
-      restDay, sumDay, daySalary, 
-      realSalary, createTime, lastModifyTime, 
-      isDelete)
-    values (#{id,jdbcType=INTEGER}, #{yearMonth,jdbcType=VARCHAR}, #{empId,jdbcType=INTEGER}, 
-      #{empName,jdbcType=VARCHAR}, #{deptId,jdbcType=INTEGER}, #{deptName,jdbcType=VARCHAR}, 
-      #{workName,jdbcType=VARCHAR}, #{salary,jdbcType=DECIMAL}, #{workDay,jdbcType=INTEGER}, 
-      #{dayWork,jdbcType=INTEGER}, #{nightWork,jdbcType=INTEGER}, #{score,jdbcType=DECIMAL}, 
-      #{restDay,jdbcType=INTEGER}, #{sumDay,jdbcType=INTEGER}, #{daySalary,jdbcType=DECIMAL}, 
-      #{realSalary,jdbcType=DECIMAL}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
-      #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_emp_salary_monthly (yearMonth, empId, empName, 
+      deptId, deptName, workName, 
+      salary, workDay, dayWork, 
+      nightWork, score, restDay, 
+      sumDay, daySalary, realSalary, 
+      createTime, lastModifyTime, isDelete
+      )
+    values (#{yearMonth,jdbcType=VARCHAR}, #{empId,jdbcType=INTEGER}, #{empName,jdbcType=VARCHAR}, 
+      #{deptId,jdbcType=INTEGER}, #{deptName,jdbcType=VARCHAR}, #{workName,jdbcType=VARCHAR}, 
+      #{salary,jdbcType=DECIMAL}, #{workDay,jdbcType=INTEGER}, #{dayWork,jdbcType=INTEGER}, 
+      #{nightWork,jdbcType=INTEGER}, #{score,jdbcType=DECIMAL}, #{restDay,jdbcType=INTEGER}, 
+      #{sumDay,jdbcType=INTEGER}, #{daySalary,jdbcType=DECIMAL}, #{realSalary,jdbcType=DECIMAL}, 
+      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
+      )
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.emp.EmpSalaryMonthly">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_emp_salary_monthly
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="yearMonth != null">
         yearMonth,
       </if>
@@ -233,9 +236,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="yearMonth != null">
         #{yearMonth,jdbcType=VARCHAR},
       </if>
