+++ src/main/java/com/btjf/mapper/emp/EmpSubsibyMonthlyMapper.java
@@ -1,18 +1,18 @@
 package com.btjf.mapper.emp;
 
-import com.btjf.model.emp.empSubsidyMonthly;
-import com.btjf.model.emp.empSubsidyMonthlyExample;
+import com.btjf.model.emp.EmpSubsibyMonthly;
+import com.btjf.model.emp.EmpSubsibyMonthlyExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-public interface empSubsidyMonthlyMapper {
+public interface EmpSubsibyMonthlyMapper {
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
-    long countByExample(empSubsidyMonthlyExample example);
+    long countByExample(EmpSubsibyMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -20,7 +20,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    int deleteByExample(empSubsidyMonthlyExample example);
+    int deleteByExample(EmpSubsibyMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -36,7 +36,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    int insert(empSubsidyMonthly record);
+    int insert(EmpSubsibyMonthly record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -44,7 +44,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    int insertSelective(empSubsidyMonthly record);
+    int insertSelective(EmpSubsibyMonthly record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -52,7 +52,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    List<empSubsidyMonthly> selectByExample(empSubsidyMonthlyExample example);
+    List<EmpSubsibyMonthly> selectByExample(EmpSubsibyMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -60,7 +60,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    empSubsidyMonthly selectByPrimaryKey(Integer id);
+    EmpSubsibyMonthly selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -68,7 +68,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByExampleSelective(@Param("record") empSubsidyMonthly record, @Param("example") empSubsidyMonthlyExample example);
+    int updateByExampleSelective(@Param("record") EmpSubsibyMonthly record, @Param("example") EmpSubsibyMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -76,7 +76,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByExample(@Param("record") empSubsidyMonthly record, @Param("example") empSubsidyMonthlyExample example);
+    int updateByExample(@Param("record") EmpSubsibyMonthly record, @Param("example") EmpSubsibyMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -84,7 +84,7 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKeySelective(empSubsidyMonthly record);
+    int updateByPrimaryKeySelective(EmpSubsibyMonthly record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -92,5 +92,5 @@ public interface empSubsidyMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKey(empSubsidyMonthly record);
+    int updateByPrimaryKey(EmpSubsibyMonthly record);
 }
