+++ src/main/java/com/btjf/mapper/emp/EmpTimesalaryMonthlyMapper.java
@@ -1,18 +1,18 @@
 package com.btjf.mapper.emp;
 
-import com.btjf.model.emp.empTimesalaryMonthly;
-import com.btjf.model.emp.empTimesalaryMonthlyExample;
+import com.btjf.model.emp.EmpTimesalaryMonthly;
+import com.btjf.model.emp.EmpTimesalaryMonthlyExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-public interface empTimesalaryMonthlyMapper {
+public interface EmpTimesalaryMonthlyMapper {
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
-    long countByExample(empTimesalaryMonthlyExample example);
+    long countByExample(EmpTimesalaryMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -20,7 +20,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    int deleteByExample(empTimesalaryMonthlyExample example);
+    int deleteByExample(EmpTimesalaryMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -36,7 +36,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    int insert(empTimesalaryMonthly record);
+    int insert(EmpTimesalaryMonthly record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -44,7 +44,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    int insertSelective(empTimesalaryMonthly record);
+    int insertSelective(EmpTimesalaryMonthly record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -52,7 +52,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    List<empTimesalaryMonthly> selectByExample(empTimesalaryMonthlyExample example);
+    List<EmpTimesalaryMonthly> selectByExample(EmpTimesalaryMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -60,7 +60,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    empTimesalaryMonthly selectByPrimaryKey(Integer id);
+    EmpTimesalaryMonthly selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -68,7 +68,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByExampleSelective(@Param("record") empTimesalaryMonthly record, @Param("example") empTimesalaryMonthlyExample example);
+    int updateByExampleSelective(@Param("record") EmpTimesalaryMonthly record, @Param("example") EmpTimesalaryMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -76,7 +76,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByExample(@Param("record") empTimesalaryMonthly record, @Param("example") empTimesalaryMonthlyExample example);
+    int updateByExample(@Param("record") EmpTimesalaryMonthly record, @Param("example") EmpTimesalaryMonthlyExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -84,7 +84,7 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKeySelective(empTimesalaryMonthly record);
+    int updateByPrimaryKeySelective(EmpTimesalaryMonthly record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -92,5 +92,5 @@ public interface empTimesalaryMonthlyMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKey(empTimesalaryMonthly record);
+    int updateByPrimaryKey(EmpTimesalaryMonthly record);
 }
