+++ src/main/java/com/btjf/mapper/emp/EmpWorkMapper.java
@@ -1,18 +1,18 @@
 package com.btjf.mapper.emp;
 
-import com.btjf.model.emp.Empwork;
-import com.btjf.model.emp.EmpworkExample;
+import com.btjf.model.emp.EmpWork;
+import com.btjf.model.emp.EmpWorkExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-public interface EmpworkMapper {
+public interface EmpWorkMapper {
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table t_emp_work
      *
      * @mbg.generated
      */
-    long countByExample(EmpworkExample example);
+    long countByExample(EmpWorkExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -20,7 +20,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    int deleteByExample(EmpworkExample example);
+    int deleteByExample(EmpWorkExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -36,7 +36,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    int insert(Empwork record);
+    int insert(EmpWork record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -44,7 +44,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    int insertSelective(Empwork record);
+    int insertSelective(EmpWork record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -52,7 +52,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    List<Empwork> selectByExample(EmpworkExample example);
+    List<EmpWork> selectByExample(EmpWorkExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -60,7 +60,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    Empwork selectByPrimaryKey(Integer id);
+    EmpWork selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -68,7 +68,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    int updateByExampleSelective(@Param("record") Empwork record, @Param("example") EmpworkExample example);
+    int updateByExampleSelective(@Param("record") EmpWork record, @Param("example") EmpWorkExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -76,7 +76,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    int updateByExample(@Param("record") Empwork record, @Param("example") EmpworkExample example);
+    int updateByExample(@Param("record") EmpWork record, @Param("example") EmpWorkExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -84,7 +84,7 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKeySelective(Empwork record);
+    int updateByPrimaryKeySelective(EmpWork record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -92,5 +92,5 @@ public interface EmpworkMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKey(Empwork record);
+    int updateByPrimaryKey(EmpWork record);
 }
