+++ src/main/java/com/btjf/mapper/emp/EmpWorkMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.btjf.mapper.emp.EmpworkMapper">
-  <resultMap id="BaseResultMap" type="com.btjf.model.emp.Empwork">
+<mapper namespace="com.btjf.mapper.emp.EmpWorkMapper">
+  <resultMap id="BaseResultMap" type="com.btjf.model.emp.EmpWork">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -85,7 +85,7 @@
     -->
     id, name, createTime, lastModifyTime, isDelete
   </sql>
-  <select id="selectByExample" parameterType="com.btjf.model.emp.EmpworkExample" resultMap="BaseResultMap">
+  <select id="selectByExample" parameterType="com.btjf.model.emp.EmpWorkExample" resultMap="BaseResultMap">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -121,7 +121,7 @@
     delete from t_emp_work
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="com.btjf.model.emp.EmpworkExample">
+  <delete id="deleteByExample" parameterType="com.btjf.model.emp.EmpWorkExample">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -131,26 +131,29 @@
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.btjf.model.emp.Empwork">
+  <insert id="insert" parameterType="com.btjf.model.emp.EmpWork">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_emp_work (id, name, createTime, 
-      lastModifyTime, isDelete)
-    values (#{id,jdbcType=INTEGER}, #{name,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, 
-      #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_emp_work (name, createTime, lastModifyTime, 
+      isDelete)
+    values (#{name,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
+      #{isDelete,jdbcType=INTEGER})
   </insert>
-  <insert id="insertSelective" parameterType="com.btjf.model.emp.Empwork">
+  <insert id="insertSelective" parameterType="com.btjf.model.emp.EmpWork">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_emp_work
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="name != null">
         name,
       </if>
@@ -165,9 +168,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="name != null">
         #{name,jdbcType=VARCHAR},
       </if>
@@ -182,7 +182,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="com.btjf.model.emp.EmpworkExample" resultType="java.lang.Long">
+  <select id="countByExample" parameterType="com.btjf.model.emp.EmpWorkExample" resultType="java.lang.Long">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -234,7 +234,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.emp.Empwork">
+  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.emp.EmpWork">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -256,7 +256,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.btjf.model.emp.Empwork">
+  <update id="updateByPrimaryKey" parameterType="com.btjf.model.emp.EmpWork">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
