+++ src/main/java/com/btjf/mapper/order/OrderProductMapper.xml
@@ -153,31 +153,34 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_order_product (id, orderId, orderNo, 
-      productId, productNo, productType, 
-      num, maxNum, unit, 
-      completeDate, customerId, customerName, 
-      isMore, urgentLevel, leader, 
-      pmType, pmCheckItem, createTime, 
-      lastModifyTime, isDelete)
-    values (#{id,jdbcType=INTEGER}, #{orderId,jdbcType=INTEGER}, #{orderNo,jdbcType=VARCHAR}, 
-      #{productId,jdbcType=INTEGER}, #{productNo,jdbcType=VARCHAR}, #{productType,jdbcType=VARCHAR}, 
-      #{num,jdbcType=INTEGER}, #{maxNum,jdbcType=INTEGER}, #{unit,jdbcType=VARCHAR}, 
-      #{completeDate,jdbcType=DATE}, #{customerId,jdbcType=INTEGER}, #{customerName,jdbcType=VARCHAR}, 
-      #{isMore,jdbcType=INTEGER}, #{urgentLevel,jdbcType=INTEGER}, #{leader,jdbcType=VARCHAR}, 
-      #{pmType,jdbcType=VARCHAR}, #{pmCheckItem,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, 
-      #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_order_product (orderId, orderNo, productId, 
+      productNo, productType, num, 
+      maxNum, unit, completeDate, 
+      customerId, customerName, isMore, 
+      urgentLevel, leader, pmType, 
+      pmCheckItem, createTime, lastModifyTime, 
+      isDelete)
+    values (#{orderId,jdbcType=INTEGER}, #{orderNo,jdbcType=VARCHAR}, #{productId,jdbcType=INTEGER}, 
+      #{productNo,jdbcType=VARCHAR}, #{productType,jdbcType=VARCHAR}, #{num,jdbcType=INTEGER}, 
+      #{maxNum,jdbcType=INTEGER}, #{unit,jdbcType=VARCHAR}, #{completeDate,jdbcType=DATE}, 
+      #{customerId,jdbcType=INTEGER}, #{customerName,jdbcType=VARCHAR}, #{isMore,jdbcType=INTEGER}, 
+      #{urgentLevel,jdbcType=INTEGER}, #{leader,jdbcType=VARCHAR}, #{pmType,jdbcType=VARCHAR}, 
+      #{pmCheckItem,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
+      #{isDelete,jdbcType=INTEGER})
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.order.OrderProduct">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_order_product
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="orderId != null">
         orderId,
       </if>
@@ -237,9 +240,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="orderId != null">
         #{orderId,jdbcType=INTEGER},
       </if>
