+++ src/main/java/com/btjf/mapper/order/OrderProductWorkshopMapper.xml
@@ -138,23 +138,26 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_order_product_workshop (id, orderProductId, workshop, 
-      leader, createTime, lastModifyTime, 
-      isDelete)
-    values (#{id,jdbcType=INTEGER}, #{orderProductId,jdbcType=INTEGER}, #{workshop,jdbcType=VARCHAR}, 
-      #{leader,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
-      #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_order_product_workshop (orderProductId, workshop, leader, 
+      createTime, lastModifyTime, isDelete
+      )
+    values (#{orderProductId,jdbcType=INTEGER}, #{workshop,jdbcType=VARCHAR}, #{leader,jdbcType=VARCHAR}, 
+      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
+      )
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.order.OrderProductWorkshop">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_order_product_workshop
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="orderProductId != null">
         orderProductId,
       </if>
@@ -175,9 +178,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="orderProductId != null">
         #{orderProductId,jdbcType=INTEGER},
       </if>
