+++ src/main/java/com/btjf/mapper/pm/PmInMapper.xml
@@ -147,29 +147,30 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_pmin (id, pmId, pmNo, 
-      pmName, type, unit, 
-      remark, supplier, num, 
-      perNum, backNum, operator, 
-      createTime, lastModifyTime, isDelete
-      )
-    values (#{id,jdbcType=INTEGER}, #{pmId,jdbcType=INTEGER}, #{pmNo,jdbcType=VARCHAR}, 
-      #{pmName,jdbcType=VARCHAR}, #{type,jdbcType=VARCHAR}, #{unit,jdbcType=VARCHAR}, 
-      #{remark,jdbcType=VARCHAR}, #{supplier,jdbcType=VARCHAR}, #{num,jdbcType=INTEGER}, 
-      #{perNum,jdbcType=INTEGER}, #{backNum,jdbcType=INTEGER}, #{operator,jdbcType=VARCHAR}, 
-      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
-      )
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_pmin (pmId, pmNo, pmName, 
+      type, unit, remark, 
+      supplier, num, perNum, 
+      backNum, operator, createTime, 
+      lastModifyTime, isDelete)
+    values (#{pmId,jdbcType=INTEGER}, #{pmNo,jdbcType=VARCHAR}, #{pmName,jdbcType=VARCHAR}, 
+      #{type,jdbcType=VARCHAR}, #{unit,jdbcType=VARCHAR}, #{remark,jdbcType=VARCHAR}, 
+      #{supplier,jdbcType=VARCHAR}, #{num,jdbcType=INTEGER}, #{perNum,jdbcType=INTEGER}, 
+      #{backNum,jdbcType=INTEGER}, #{operator,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, 
+      #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER})
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.pm.PmIn">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_pmin
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="pmId != null">
         pmId,
       </if>
@@ -214,9 +215,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="pmId != null">
         #{pmId,jdbcType=INTEGER},
       </if>
