+++ src/main/java/com/btjf/mapper/pm/PmMapper.java
@@ -95,7 +95,6 @@ public interface PmMapper {
      */
     int updateByPrimaryKey(Pm record);
 
-
     List<Pm> findList(@Param("pmNo") String pmNo, @Param("name") String name,@Param("type") String type);
 
     Integer deleteByID(@Param("integers") List<Integer> integers);
