+++ src/main/java/com/btjf/mapper/pm/PmOutBillDetailMapper.java
@@ -1,18 +1,19 @@
 package com.btjf.mapper.pm;
 
-import com.btjf.model.pm.PmoutBillDetail;
-import com.btjf.model.pm.PmoutBillDetailExample;
-import java.util.List;
+import com.btjf.model.pm.PmOutBillDetail;
+import com.btjf.model.pm.PmOutBillDetailExample;
 import org.apache.ibatis.annotations.Param;
 
-public interface PmoutBillDetailMapper {
+import java.util.List;
+
+public interface PmOutBillDetailMapper {
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
-    long countByExample(PmoutBillDetailExample example);
+    long countByExample(PmOutBillDetailExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -20,7 +21,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    int deleteByExample(PmoutBillDetailExample example);
+    int deleteByExample(PmOutBillDetailExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -36,7 +37,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    int insert(PmoutBillDetail record);
+    int insert(PmOutBillDetail record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -44,7 +45,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    int insertSelective(PmoutBillDetail record);
+    int insertSelective(PmOutBillDetail record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -52,7 +53,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    List<PmoutBillDetail> selectByExample(PmoutBillDetailExample example);
+    List<PmOutBillDetail> selectByExample(PmOutBillDetailExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -60,7 +61,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    PmoutBillDetail selectByPrimaryKey(Integer id);
+    PmOutBillDetail selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -68,7 +69,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    int updateByExampleSelective(@Param("record") PmoutBillDetail record, @Param("example") PmoutBillDetailExample example);
+    int updateByExampleSelective(@Param("record") PmOutBillDetail record, @Param("example") PmOutBillDetailExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -76,7 +77,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    int updateByExample(@Param("record") PmoutBillDetail record, @Param("example") PmoutBillDetailExample example);
+    int updateByExample(@Param("record") PmOutBillDetail record, @Param("example") PmOutBillDetailExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -84,7 +85,7 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKeySelective(PmoutBillDetail record);
+    int updateByPrimaryKeySelective(PmOutBillDetail record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -92,5 +93,5 @@ public interface PmoutBillDetailMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKey(PmoutBillDetail record);
+    int updateByPrimaryKey(PmOutBillDetail record);
 }
