+++ src/main/java/com/btjf/mapper/pm/PmOutBillDetailMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.btjf.mapper.pm.PmoutBillDetailMapper">
-  <resultMap id="BaseResultMap" type="com.btjf.model.pm.PmoutBillDetail">
+<mapper namespace="com.btjf.mapper.pm.PmOutBillDetailMapper">
+  <resultMap id="BaseResultMap" type="com.btjf.model.pm.PmOutBillDetail">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -99,7 +99,7 @@
     id, billId, pmId, pmNo, pmName, pmBatchNo, num, perNum, backNum, unit, remark, outInfo, 
     receiver, sender, operator, createTime, lastModifyTime, isDelete
   </sql>
-  <select id="selectByExample" parameterType="com.btjf.model.pm.PmoutBillDetailExample" resultMap="BaseResultMap">
+  <select id="selectByExample" parameterType="com.btjf.model.pm.PmOutBillDetailExample" resultMap="BaseResultMap">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -135,7 +135,7 @@
     delete from t_pmout_bill_detail
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="com.btjf.model.pm.PmoutBillDetailExample">
+  <delete id="deleteByExample" parameterType="com.btjf.model.pm.PmOutBillDetailExample">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -145,36 +145,37 @@
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.btjf.model.pm.PmoutBillDetail">
+  <insert id="insert" parameterType="com.btjf.model.pm.PmOutBillDetail">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_pmout_bill_detail (id, billId, pmId, 
-      pmNo, pmName, pmBatchNo, 
-      num, perNum, backNum, 
-      unit, remark, outInfo, 
-      receiver, sender, operator, 
-      createTime, lastModifyTime, isDelete
-      )
-    values (#{id,jdbcType=INTEGER}, #{billId,jdbcType=INTEGER}, #{pmId,jdbcType=INTEGER}, 
-      #{pmNo,jdbcType=VARCHAR}, #{pmName,jdbcType=VARCHAR}, #{pmBatchNo,jdbcType=VARCHAR}, 
-      #{num,jdbcType=INTEGER}, #{perNum,jdbcType=INTEGER}, #{backNum,jdbcType=INTEGER}, 
-      #{unit,jdbcType=VARCHAR}, #{remark,jdbcType=VARCHAR}, #{outInfo,jdbcType=VARCHAR}, 
-      #{receiver,jdbcType=VARCHAR}, #{sender,jdbcType=VARCHAR}, #{operator,jdbcType=VARCHAR}, 
-      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
-      )
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_pmout_bill_detail (billId, pmId, pmNo, 
+      pmName, pmBatchNo, num, 
+      perNum, backNum, unit, 
+      remark, outInfo, receiver, 
+      sender, operator, createTime, 
+      lastModifyTime, isDelete)
+    values (#{billId,jdbcType=INTEGER}, #{pmId,jdbcType=INTEGER}, #{pmNo,jdbcType=VARCHAR}, 
+      #{pmName,jdbcType=VARCHAR}, #{pmBatchNo,jdbcType=VARCHAR}, #{num,jdbcType=INTEGER}, 
+      #{perNum,jdbcType=INTEGER}, #{backNum,jdbcType=INTEGER}, #{unit,jdbcType=VARCHAR}, 
+      #{remark,jdbcType=VARCHAR}, #{outInfo,jdbcType=VARCHAR}, #{receiver,jdbcType=VARCHAR}, 
+      #{sender,jdbcType=VARCHAR}, #{operator,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, 
+      #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER})
   </insert>
-  <insert id="insertSelective" parameterType="com.btjf.model.pm.PmoutBillDetail">
+  <insert id="insertSelective" parameterType="com.btjf.model.pm.PmOutBillDetail">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_pmout_bill_detail
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="billId != null">
         billId,
       </if>
@@ -228,9 +229,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="billId != null">
         #{billId,jdbcType=INTEGER},
       </if>
@@ -284,7 +282,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="com.btjf.model.pm.PmoutBillDetailExample" resultType="java.lang.Long">
+  <select id="countByExample" parameterType="com.btjf.model.pm.PmOutBillDetailExample" resultType="java.lang.Long">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -388,7 +386,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.pm.PmoutBillDetail">
+  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.pm.PmOutBillDetail">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -449,7 +447,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.btjf.model.pm.PmoutBillDetail">
+  <update id="updateByPrimaryKey" parameterType="com.btjf.model.pm.PmOutBillDetail">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
