+++ src/main/java/com/btjf/mapper/pm/PmOutBillMapper.java
@@ -1,18 +1,19 @@
 package com.btjf.mapper.pm;
 
-import com.btjf.model.pm.pmOutBill;
-import com.btjf.model.pm.pmOutBillExample;
-import java.util.List;
+import com.btjf.model.pm.PmOutBill;
+import com.btjf.model.pm.PmOutBillExample;
 import org.apache.ibatis.annotations.Param;
 
-public interface pmOutBillMapper {
+import java.util.List;
+
+public interface PmOutBillMapper {
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table t_pmout_bill
      *
      * @mbg.generated
      */
-    long countByExample(pmOutBillExample example);
+    long countByExample(PmOutBillExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -20,7 +21,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    int deleteByExample(pmOutBillExample example);
+    int deleteByExample(PmOutBillExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -36,7 +37,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    int insert(pmOutBill record);
+    int insert(PmOutBill record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -44,7 +45,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    int insertSelective(pmOutBill record);
+    int insertSelective(PmOutBill record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -52,7 +53,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    List<pmOutBill> selectByExample(pmOutBillExample example);
+    List<PmOutBill> selectByExample(PmOutBillExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -60,7 +61,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    pmOutBill selectByPrimaryKey(Integer id);
+    PmOutBill selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -68,7 +69,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    int updateByExampleSelective(@Param("record") pmOutBill record, @Param("example") pmOutBillExample example);
+    int updateByExampleSelective(@Param("record") PmOutBill record, @Param("example") PmOutBillExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -76,7 +77,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    int updateByExample(@Param("record") pmOutBill record, @Param("example") pmOutBillExample example);
+    int updateByExample(@Param("record") PmOutBill record, @Param("example") PmOutBillExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -84,7 +85,7 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKeySelective(pmOutBill record);
+    int updateByPrimaryKeySelective(PmOutBill record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -92,5 +93,5 @@ public interface pmOutBillMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKey(pmOutBill record);
+    int updateByPrimaryKey(PmOutBill record);
 }
