+++ src/main/java/com/btjf/mapper/pm/PmOutBillMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.btjf.mapper.pm.pmOutBillMapper">
-  <resultMap id="BaseResultMap" type="com.btjf.model.pm.pmOutBill">
+<mapper namespace="com.btjf.mapper.pm.PmOutBillMapper">
+  <resultMap id="BaseResultMap" type="com.btjf.model.pm.PmOutBill">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -97,7 +97,7 @@
     id, orderId, orderNo, productId, productNo, billNo, type, workshop, distributionNum, 
     groupNmae, leader, pmType, pmCheckItem, createTime, lastModifyTime, isDelete
   </sql>
-  <select id="selectByExample" parameterType="com.btjf.model.pm.pmOutBillExample" resultMap="BaseResultMap">
+  <select id="selectByExample" parameterType="com.btjf.model.pm.PmOutBillExample" resultMap="BaseResultMap">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -133,7 +133,7 @@
     delete from t_pmout_bill
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="com.btjf.model.pm.pmOutBillExample">
+  <delete id="deleteByExample" parameterType="com.btjf.model.pm.PmOutBillExample">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -143,34 +143,37 @@
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.btjf.model.pm.pmOutBill">
+  <insert id="insert" parameterType="com.btjf.model.pm.PmOutBill">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_pmout_bill (id, orderId, orderNo, 
-      productId, productNo, billNo, 
-      type, workshop, distributionNum, 
-      groupNmae, leader, pmType, 
-      pmCheckItem, createTime, lastModifyTime, 
-      isDelete)
-    values (#{id,jdbcType=INTEGER}, #{orderId,jdbcType=INTEGER}, #{orderNo,jdbcType=VARCHAR}, 
-      #{productId,jdbcType=INTEGER}, #{productNo,jdbcType=VARCHAR}, #{billNo,jdbcType=VARCHAR}, 
-      #{type,jdbcType=INTEGER}, #{workshop,jdbcType=INTEGER}, #{distributionNum,jdbcType=INTEGER}, 
-      #{groupNmae,jdbcType=VARCHAR}, #{leader,jdbcType=VARCHAR}, #{pmType,jdbcType=VARCHAR}, 
-      #{pmCheckItem,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
-      #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_pmout_bill (orderId, orderNo, productId, 
+      productNo, billNo, type, 
+      workshop, distributionNum, groupNmae, 
+      leader, pmType, pmCheckItem, 
+      createTime, lastModifyTime, isDelete
+      )
+    values (#{orderId,jdbcType=INTEGER}, #{orderNo,jdbcType=VARCHAR}, #{productId,jdbcType=INTEGER}, 
+      #{productNo,jdbcType=VARCHAR}, #{billNo,jdbcType=VARCHAR}, #{type,jdbcType=INTEGER}, 
+      #{workshop,jdbcType=INTEGER}, #{distributionNum,jdbcType=INTEGER}, #{groupNmae,jdbcType=VARCHAR}, 
+      #{leader,jdbcType=VARCHAR}, #{pmType,jdbcType=VARCHAR}, #{pmCheckItem,jdbcType=VARCHAR}, 
+      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
+      )
   </insert>
-  <insert id="insertSelective" parameterType="com.btjf.model.pm.pmOutBill">
+  <insert id="insertSelective" parameterType="com.btjf.model.pm.PmOutBill">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_pmout_bill
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="orderId != null">
         orderId,
       </if>
@@ -218,9 +221,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="orderId != null">
         #{orderId,jdbcType=INTEGER},
       </if>
@@ -268,7 +268,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="com.btjf.model.pm.pmOutBillExample" resultType="java.lang.Long">
+  <select id="countByExample" parameterType="com.btjf.model.pm.PmOutBillExample" resultType="java.lang.Long">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -364,7 +364,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.pm.pmOutBill">
+  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.pm.PmOutBill">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -419,7 +419,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.btjf.model.pm.pmOutBill">
+  <update id="updateByPrimaryKey" parameterType="com.btjf.model.pm.PmOutBill">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
