+++ src/main/java/com/btjf/mapper/product/ProductPmMapper.java
@@ -1,19 +1,19 @@
 package com.btjf.mapper.product;
 
-import com.btjf.model.product.Productpm;
-import com.btjf.model.product.ProductpmExample;
+import com.btjf.model.product.ProductPm;
+import com.btjf.model.product.ProductPmExample;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 
-public interface ProductpmMapper {
+public interface ProductPmMapper {
     /**
      * This method was generated by MyBatis Generator.
      * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
-    long countByExample(ProductpmExample example);
+    long countByExample(ProductPmExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -21,7 +21,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    int deleteByExample(ProductpmExample example);
+    int deleteByExample(ProductPmExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -37,7 +37,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    int insert(Productpm record);
+    int insert(ProductPm record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -45,7 +45,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    int insertSelective(Productpm record);
+    int insertSelective(ProductPm record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -53,7 +53,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    List<Productpm> selectByExample(ProductpmExample example);
+    List<ProductPm> selectByExample(ProductPmExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -61,7 +61,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    Productpm selectByPrimaryKey(Integer id);
+    ProductPm selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -69,7 +69,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    int updateByExampleSelective(@Param("record") Productpm record, @Param("example") ProductpmExample example);
+    int updateByExampleSelective(@Param("record") ProductPm record, @Param("example") ProductPmExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -77,7 +77,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    int updateByExample(@Param("record") Productpm record, @Param("example") ProductpmExample example);
+    int updateByExample(@Param("record") ProductPm record, @Param("example") ProductPmExample example);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -85,7 +85,7 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKeySelective(Productpm record);
+    int updateByPrimaryKeySelective(ProductPm record);
 
     /**
      * This method was generated by MyBatis Generator.
@@ -93,7 +93,9 @@ public interface ProductpmMapper {
      *
      * @mbg.generated
      */
-    int updateByPrimaryKey(Productpm record);
+    int updateByPrimaryKey(ProductPm record);
 
-    List<Productpm> findList(@Param("productNo") String productNo, @Param("pmNo") String pmNo, @Param("status") int status);
+    List<ProductPm> findList(@Param("productNo") String productNo, @Param("pmNo") String pmNo, @Param("status") int status);
+
+    ProductPm selectByNo(@Param("productNo") String productNo);
 }
