+++ src/main/java/com/btjf/mapper/product/ProductProcedureMapper.xml
@@ -142,25 +142,28 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_product_procedure (id, procedureName, productId, 
-      productNo, sort, price, 
-      operator, createTime, lastModifyTime, 
-      isDelete)
-    values (#{id,jdbcType=INTEGER}, #{procedureName,jdbcType=VARCHAR}, #{productId,jdbcType=INTEGER}, 
-      #{productNo,jdbcType=VARCHAR}, #{sort,jdbcType=INTEGER}, #{price,jdbcType=DECIMAL}, 
-      #{operator,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
-      #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_product_procedure (procedureName, productId, productNo, 
+      sort, price, operator, 
+      createTime, lastModifyTime, isDelete
+      )
+    values (#{procedureName,jdbcType=VARCHAR}, #{productId,jdbcType=INTEGER}, #{productNo,jdbcType=VARCHAR}, 
+      #{sort,jdbcType=INTEGER}, #{price,jdbcType=DECIMAL}, #{operator,jdbcType=VARCHAR}, 
+      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
+      )
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.product.ProductProcedure">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_product_procedure
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="procedureName != null">
         procedureName,
       </if>
@@ -190,9 +193,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="procedureName != null">
         #{procedureName,jdbcType=VARCHAR},
       </if>
