+++ src/main/java/com/btjf/mapper/product/ProductProcedureWorkshopMapper.xml
@@ -145,27 +145,30 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_product_procedure_workshop (id, productId, productNo, 
-      sort, workshop, procedureId, 
-      procedureName, price, remark, 
-      operator, createTime, lastModifyTime, 
-      isDelete)
-    values (#{id,jdbcType=INTEGER}, #{productId,jdbcType=INTEGER}, #{productNo,jdbcType=VARCHAR}, 
-      #{sort,jdbcType=INTEGER}, #{workshop,jdbcType=VARCHAR}, #{procedureId,jdbcType=INTEGER}, 
-      #{procedureName,jdbcType=VARCHAR}, #{price,jdbcType=DECIMAL}, #{remark,jdbcType=VARCHAR}, 
-      #{operator,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
-      #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_product_procedure_workshop (productId, productNo, sort, 
+      workshop, procedureId, procedureName, 
+      price, remark, operator, 
+      createTime, lastModifyTime, isDelete
+      )
+    values (#{productId,jdbcType=INTEGER}, #{productNo,jdbcType=VARCHAR}, #{sort,jdbcType=INTEGER}, 
+      #{workshop,jdbcType=VARCHAR}, #{procedureId,jdbcType=INTEGER}, #{procedureName,jdbcType=VARCHAR}, 
+      #{price,jdbcType=DECIMAL}, #{remark,jdbcType=VARCHAR}, #{operator,jdbcType=VARCHAR}, 
+      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
+      )
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.product.ProductProcedureWorkshop">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_product_procedure_workshop
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="productId != null">
         productId,
       </if>
@@ -204,9 +207,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="productId != null">
         #{productId,jdbcType=INTEGER},
       </if>
