+++ src/main/java/com/btjf/mapper/product/ProductpmMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.btjf.mapper.product.ProductpmMapper">
-  <resultMap id="BaseResultMap" type="com.btjf.model.product.Productpm">
+<mapper namespace="com.btjf.mapper.product.ProductPmMapper">
+  <resultMap id="BaseResultMap" type="com.btjf.model.product.ProductPm">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -97,7 +97,7 @@
     id, productId, productNo, pmId, pmNo, pmName, type, unit, remark, status, num, unitNum, 
     operator, createTime, lastModifyTime, isDelete
   </sql>
-  <select id="selectByExample" parameterType="com.btjf.model.product.ProductpmExample" resultMap="BaseResultMap">
+  <select id="selectByExample" parameterType="com.btjf.model.product.ProductPmExample" resultMap="BaseResultMap">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -133,7 +133,7 @@
     delete from t_productpm
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="com.btjf.model.product.ProductpmExample">
+  <delete id="deleteByExample" parameterType="com.btjf.model.product.ProductPmExample">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -143,34 +143,37 @@
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.btjf.model.product.Productpm">
+  <insert id="insert" parameterType="com.btjf.model.product.ProductPm">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_productpm (id, productId, productNo, 
-      pmId, pmNo, pmName, 
-      type, unit, remark, 
-      status, num, unitNum, 
-      operator, createTime, lastModifyTime, 
-      isDelete)
-    values (#{id,jdbcType=INTEGER}, #{productId,jdbcType=INTEGER}, #{productNo,jdbcType=VARCHAR}, 
-      #{pmId,jdbcType=INTEGER}, #{pmNo,jdbcType=VARCHAR}, #{pmName,jdbcType=VARCHAR}, 
-      #{type,jdbcType=VARCHAR}, #{unit,jdbcType=VARCHAR}, #{remark,jdbcType=VARCHAR}, 
-      #{status,jdbcType=INTEGER}, #{num,jdbcType=DECIMAL}, #{unitNum,jdbcType=DECIMAL}, 
-      #{operator,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
-      #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_productpm (productId, productNo, pmId, 
+      pmNo, pmName, type, 
+      unit, remark, status, 
+      num, unitNum, operator, 
+      createTime, lastModifyTime, isDelete
+      )
+    values (#{productId,jdbcType=INTEGER}, #{productNo,jdbcType=VARCHAR}, #{pmId,jdbcType=INTEGER}, 
+      #{pmNo,jdbcType=VARCHAR}, #{pmName,jdbcType=VARCHAR}, #{type,jdbcType=VARCHAR}, 
+      #{unit,jdbcType=VARCHAR}, #{remark,jdbcType=VARCHAR}, #{status,jdbcType=INTEGER}, 
+      #{num,jdbcType=DECIMAL}, #{unitNum,jdbcType=DECIMAL}, #{operator,jdbcType=VARCHAR}, 
+      #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER}
+      )
   </insert>
-  <insert id="insertSelective" parameterType="com.btjf.model.product.Productpm">
+  <insert id="insertSelective" parameterType="com.btjf.model.product.ProductPm">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_productpm
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="productId != null">
         productId,
       </if>
@@ -218,9 +221,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="productId != null">
         #{productId,jdbcType=INTEGER},
       </if>
@@ -268,7 +268,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="com.btjf.model.product.ProductpmExample" resultType="java.lang.Long">
+  <select id="countByExample" parameterType="com.btjf.model.product.ProductPmExample" resultType="java.lang.Long">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -364,7 +364,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.product.Productpm">
+  <update id="updateByPrimaryKeySelective" parameterType="com.btjf.model.product.ProductPm">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -419,7 +419,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.btjf.model.product.Productpm">
+  <update id="updateByPrimaryKey" parameterType="com.btjf.model.product.ProductPm">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
@@ -444,7 +444,7 @@
   </update>
 
   <select id="findList" resultMap="BaseResultMap">
-    SELECT <include refid="Base_Column_List"/>
+    SELECT <include refid="Base_Column_List" />
     FROM t_productpm
     WHERE
     1=1
@@ -462,5 +462,11 @@
     </if>
     AND isDelete = 0
     order by createTime desc
+  </select>
+
+  <select id="selectByNo" resultMap="BaseResultMap">
+    select <include refid="Base_Column_List"/>
+    from t_productpm
+    WHERE productNo = #{productNo} and isDelete = 0;
   </select>
 </mapper>
