+++ src/main/java/com/btjf/mapper/salary/SalaryMonthlyMapper.xml
@@ -140,23 +140,26 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_salary_monthly (id, yearMonth, expectWorkDay, 
-      realityWorkDay, basicSalary, createTime, 
-      lastModifyTime, isDelete)
-    values (#{id,jdbcType=INTEGER}, #{yearMonth,jdbcType=VARCHAR}, #{expectWorkDay,jdbcType=INTEGER}, 
-      #{realityWorkDay,jdbcType=INTEGER}, #{basicSalary,jdbcType=DECIMAL}, #{createTime,jdbcType=TIMESTAMP}, 
-      #{lastModifyTime,jdbcType=TIMESTAMP}, #{isDelete,jdbcType=INTEGER})
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_salary_monthly (yearMonth, expectWorkDay, realityWorkDay, 
+      basicSalary, createTime, lastModifyTime, 
+      isDelete)
+    values (#{yearMonth,jdbcType=VARCHAR}, #{expectWorkDay,jdbcType=INTEGER}, #{realityWorkDay,jdbcType=INTEGER}, 
+      #{basicSalary,jdbcType=DECIMAL}, #{createTime,jdbcType=TIMESTAMP}, #{lastModifyTime,jdbcType=TIMESTAMP}, 
+      #{isDelete,jdbcType=INTEGER})
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.salary.SalaryMonthly">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_salary_monthly
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="yearMonth != null">
         yearMonth,
       </if>
@@ -180,9 +183,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="yearMonth != null">
         #{yearMonth,jdbcType=VARCHAR},
       </if>
