+++ src/main/java/com/btjf/mapper/sys/SysUserMapper.xml
@@ -137,23 +137,24 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_sysuser (id, deptId, roleId, 
-      loginName, loginPwd, userName
-      )
-    values (#{id,jdbcType=INTEGER}, #{deptId,jdbcType=INTEGER}, #{roleId,jdbcType=INTEGER}, 
-      #{loginName,jdbcType=VARCHAR}, #{loginPwd,jdbcType=VARCHAR}, #{userName,jdbcType=VARCHAR}
-      )
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_sysuser (deptId, roleId, loginName, 
+      loginPwd, userName)
+    values (#{deptId,jdbcType=INTEGER}, #{roleId,jdbcType=INTEGER}, #{loginName,jdbcType=VARCHAR}, 
+      #{loginPwd,jdbcType=VARCHAR}, #{userName,jdbcType=VARCHAR})
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.sys.SysUser">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_sysuser
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="deptId != null">
         deptId,
       </if>
@@ -171,9 +172,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="deptId != null">
         #{deptId,jdbcType=INTEGER},
       </if>
