+++ src/main/java/com/btjf/mapper/sys/SysdeptMapper.xml
@@ -134,21 +134,22 @@
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into t_sysdept (id, deptName, deptOrder
-      )
-    values (#{id,jdbcType=INTEGER}, #{deptName,jdbcType=VARCHAR}, #{deptOrder,jdbcType=INTEGER}
-      )
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
+    insert into t_sysdept (deptName, deptOrder)
+    values (#{deptName,jdbcType=VARCHAR}, #{deptOrder,jdbcType=INTEGER})
   </insert>
   <insert id="insertSelective" parameterType="com.btjf.model.sys.Sysdept">
     <!--
       WARNING - @mbg.generated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <selectKey keyProperty="id" order="AFTER" resultType="java.lang.Integer">
+      SELECT LAST_INSERT_ID()
+    </selectKey>
     insert into t_sysdept
     <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
       <if test="deptName != null">
         deptName,
       </if>
@@ -157,9 +158,6 @@
       </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=INTEGER},
-      </if>
       <if test="deptName != null">
         #{deptName,jdbcType=VARCHAR},
       </if>
