+++ src/main/java/com/btjf/model/emp/Emp.java
@@ -71,11 +71,11 @@ public class Emp implements Serializable {
     /**
      *
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_emp.native
+     * This field corresponds to the database column t_emp.nativeSource
      *
      * @mbg.generated
      */
-    //private String native;
+    private String nativeSource;
 
     /**
      *
@@ -471,27 +471,27 @@ public class Emp implements Serializable {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_emp.native
+     * This method returns the value of the database column t_emp.nativeSource
      *
-     * @return the value of t_emp.native
+     * @return the value of t_emp.nativeSource
      *
      * @mbg.generated
      */
-  /*  public String getNative() {
-        return native;
+    public String getNativeSource() {
+        return nativeSource;
     }
 
-    *//**
+    /**
      * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_emp.native
+     * This method sets the value of the database column t_emp.nativeSource
      *
-     * @param native the value for t_emp.native
+     * @param nativeSource the value for t_emp.nativeSource
      *
      * @mbg.generated
-     *//*
-    public void setNative(String native) {
-        this.native = native == null ? null : native.trim();
-    }*/
+     */
+    public void setNativeSource(String nativeSource) {
+        this.nativeSource = nativeSource == null ? null : nativeSource.trim();
+    }
 
     /**
      * This method was generated by MyBatis Generator.
@@ -1094,6 +1094,7 @@ public class Emp implements Serializable {
             && (this.getSex() == null ? other.getSex() == null : this.getSex().equals(other.getSex()))
             && (this.getWorkId() == null ? other.getWorkId() == null : this.getWorkId().equals(other.getWorkId()))
             && (this.getType() == null ? other.getType() == null : this.getType().equals(other.getType()))
+            && (this.getNativeSource() == null ? other.getNativeSource() == null : this.getNativeSource().equals(other.getNativeSource()))
             && (this.getNation() == null ? other.getNation() == null : this.getNation().equals(other.getNation()))
             && (this.getBirthday() == null ? other.getBirthday() == null : this.getBirthday().equals(other.getBirthday()))
             && (this.getIdCard() == null ? other.getIdCard() == null : this.getIdCard().equals(other.getIdCard()))
@@ -1137,6 +1138,7 @@ public class Emp implements Serializable {
         result = prime * result + ((getSex() == null) ? 0 : getSex().hashCode());
         result = prime * result + ((getWorkId() == null) ? 0 : getWorkId().hashCode());
         result = prime * result + ((getType() == null) ? 0 : getType().hashCode());
+        result = prime * result + ((getNativeSource() == null) ? 0 : getNativeSource().hashCode());
         result = prime * result + ((getNation() == null) ? 0 : getNation().hashCode());
         result = prime * result + ((getBirthday() == null) ? 0 : getBirthday().hashCode());
         result = prime * result + ((getIdCard() == null) ? 0 : getIdCard().hashCode());
