+++ src/main/java/com/btjf/model/emp/EmpSubsibyMonthly.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.Date;
 
-public class empSubsidyMonthly implements Serializable {
+public class EmpSubsibyMonthly implements Serializable {
     /**
      *
      * This field was generated by MyBatis Generator.
@@ -624,7 +624,7 @@ public class empSubsidyMonthly implements Serializable {
         if (getClass() != that.getClass()) {
             return false;
         }
-        empSubsidyMonthly other = (empSubsidyMonthly) that;
+        EmpSubsibyMonthly other = (EmpSubsibyMonthly) that;
         return (this.getId() == null ? other.getId() == null : this.getId().equals(other.getId()))
             && (this.getYearMonth() == null ? other.getYearMonth() == null : this.getYearMonth().equals(other.getYearMonth()))
             && (this.getEmpId() == null ? other.getEmpId() == null : this.getEmpId().equals(other.getEmpId()))
