+++ src/main/java/com/btjf/model/emp/EmpSubsibyMonthlyExample.java
@@ -5,10 +5,10 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class empTimesalaryMonthlyExample {
+public class EmpSubsibyMonthlyExample {
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_emp_timesalary_monthly
+     * This field corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -16,7 +16,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_emp_timesalary_monthly
+     * This field corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -24,7 +24,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_emp_timesalary_monthly
+     * This field corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -32,17 +32,17 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
-    public empTimesalaryMonthlyExample() {
+    public EmpSubsibyMonthlyExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -52,7 +52,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -62,7 +62,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -72,7 +72,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -82,7 +82,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -92,7 +92,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -102,7 +102,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -114,7 +114,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -128,7 +128,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -139,7 +139,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_timesalary_monthly
+     * This method corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -151,7 +151,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_emp_timesalary_monthly
+     * This class corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
@@ -656,123 +656,63 @@ public class empTimesalaryMonthlyExample {
             return (Criteria) this;
         }
 
-        public Criteria andNumIsNull() {
-            addCriterion("num is null");
+        public Criteria andTypeIsNull() {
+            addCriterion("type is null");
             return (Criteria) this;
         }
 
-        public Criteria andNumIsNotNull() {
-            addCriterion("num is not null");
+        public Criteria andTypeIsNotNull() {
+            addCriterion("type is not null");
             return (Criteria) this;
         }
 
-        public Criteria andNumEqualTo(BigDecimal value) {
-            addCriterion("num =", value, "num");
+        public Criteria andTypeEqualTo(Integer value) {
+            addCriterion("type =", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumNotEqualTo(BigDecimal value) {
-            addCriterion("num <>", value, "num");
+        public Criteria andTypeNotEqualTo(Integer value) {
+            addCriterion("type <>", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumGreaterThan(BigDecimal value) {
-            addCriterion("num >", value, "num");
+        public Criteria andTypeGreaterThan(Integer value) {
+            addCriterion("type >", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumGreaterThanOrEqualTo(BigDecimal value) {
-            addCriterion("num >=", value, "num");
+        public Criteria andTypeGreaterThanOrEqualTo(Integer value) {
+            addCriterion("type >=", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumLessThan(BigDecimal value) {
-            addCriterion("num <", value, "num");
+        public Criteria andTypeLessThan(Integer value) {
+            addCriterion("type <", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumLessThanOrEqualTo(BigDecimal value) {
-            addCriterion("num <=", value, "num");
+        public Criteria andTypeLessThanOrEqualTo(Integer value) {
+            addCriterion("type <=", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumIn(List<BigDecimal> values) {
-            addCriterion("num in", values, "num");
+        public Criteria andTypeIn(List<Integer> values) {
+            addCriterion("type in", values, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumNotIn(List<BigDecimal> values) {
-            addCriterion("num not in", values, "num");
+        public Criteria andTypeNotIn(List<Integer> values) {
+            addCriterion("type not in", values, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("num between", value1, value2, "num");
+        public Criteria andTypeBetween(Integer value1, Integer value2) {
+            addCriterion("type between", value1, value2, "type");
             return (Criteria) this;
         }
 
-        public Criteria andNumNotBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("num not between", value1, value2, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceIsNull() {
-            addCriterion("price is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceIsNotNull() {
-            addCriterion("price is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceEqualTo(BigDecimal value) {
-            addCriterion("price =", value, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceNotEqualTo(BigDecimal value) {
-            addCriterion("price <>", value, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceGreaterThan(BigDecimal value) {
-            addCriterion("price >", value, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceGreaterThanOrEqualTo(BigDecimal value) {
-            addCriterion("price >=", value, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceLessThan(BigDecimal value) {
-            addCriterion("price <", value, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceLessThanOrEqualTo(BigDecimal value) {
-            addCriterion("price <=", value, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceIn(List<BigDecimal> values) {
-            addCriterion("price in", values, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceNotIn(List<BigDecimal> values) {
-            addCriterion("price not in", values, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("price between", value1, value2, "price");
-            return (Criteria) this;
-        }
-
-        public Criteria andPriceNotBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("price not between", value1, value2, "price");
+        public Criteria andTypeNotBetween(Integer value1, Integer value2) {
+            addCriterion("type not between", value1, value2, "type");
             return (Criteria) this;
         }
 
@@ -906,73 +846,73 @@ public class empTimesalaryMonthlyExample {
             return (Criteria) this;
         }
 
-        public Criteria andContentIsNull() {
-            addCriterion("content is null");
+        public Criteria andReasonIsNull() {
+            addCriterion("reason is null");
             return (Criteria) this;
         }
 
-        public Criteria andContentIsNotNull() {
-            addCriterion("content is not null");
+        public Criteria andReasonIsNotNull() {
+            addCriterion("reason is not null");
             return (Criteria) this;
         }
 
-        public Criteria andContentEqualTo(String value) {
-            addCriterion("content =", value, "content");
+        public Criteria andReasonEqualTo(String value) {
+            addCriterion("reason =", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentNotEqualTo(String value) {
-            addCriterion("content <>", value, "content");
+        public Criteria andReasonNotEqualTo(String value) {
+            addCriterion("reason <>", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentGreaterThan(String value) {
-            addCriterion("content >", value, "content");
+        public Criteria andReasonGreaterThan(String value) {
+            addCriterion("reason >", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentGreaterThanOrEqualTo(String value) {
-            addCriterion("content >=", value, "content");
+        public Criteria andReasonGreaterThanOrEqualTo(String value) {
+            addCriterion("reason >=", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentLessThan(String value) {
-            addCriterion("content <", value, "content");
+        public Criteria andReasonLessThan(String value) {
+            addCriterion("reason <", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentLessThanOrEqualTo(String value) {
-            addCriterion("content <=", value, "content");
+        public Criteria andReasonLessThanOrEqualTo(String value) {
+            addCriterion("reason <=", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentLike(String value) {
-            addCriterion("content like", value, "content");
+        public Criteria andReasonLike(String value) {
+            addCriterion("reason like", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentNotLike(String value) {
-            addCriterion("content not like", value, "content");
+        public Criteria andReasonNotLike(String value) {
+            addCriterion("reason not like", value, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentIn(List<String> values) {
-            addCriterion("content in", values, "content");
+        public Criteria andReasonIn(List<String> values) {
+            addCriterion("reason in", values, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentNotIn(List<String> values) {
-            addCriterion("content not in", values, "content");
+        public Criteria andReasonNotIn(List<String> values) {
+            addCriterion("reason not in", values, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentBetween(String value1, String value2) {
-            addCriterion("content between", value1, value2, "content");
+        public Criteria andReasonBetween(String value1, String value2) {
+            addCriterion("reason between", value1, value2, "reason");
             return (Criteria) this;
         }
 
-        public Criteria andContentNotBetween(String value1, String value2) {
-            addCriterion("content not between", value1, value2, "content");
+        public Criteria andReasonNotBetween(String value1, String value2) {
+            addCriterion("reason not between", value1, value2, "reason");
             return (Criteria) this;
         }
 
@@ -1441,8 +1381,8 @@ public class empTimesalaryMonthlyExample {
             return (Criteria) this;
         }
 
-        public Criteria andContentLikeInsensitive(String value) {
-            addCriterion("upper(content) like", value.toUpperCase(), "content");
+        public Criteria andReasonLikeInsensitive(String value) {
+            addCriterion("upper(reason) like", value.toUpperCase(), "reason");
             return (Criteria) this;
         }
 
@@ -1459,7 +1399,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_emp_timesalary_monthly
+     * This class corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated do_not_delete_during_merge
      */
@@ -1472,7 +1412,7 @@ public class empTimesalaryMonthlyExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_emp_timesalary_monthly
+     * This class corresponds to the database table t_emp_subsidy_monthly
      *
      * @mbg.generated
      */
