+++ src/main/java/com/btjf/model/emp/EmpTimesalaryMonthly.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.Date;
 
-public class empTimesalaryMonthly implements Serializable {
+public class EmpTimesalaryMonthly implements Serializable {
     /**
      *
      * This field was generated by MyBatis Generator.
@@ -657,7 +657,7 @@ public class empTimesalaryMonthly implements Serializable {
         if (getClass() != that.getClass()) {
             return false;
         }
-        empTimesalaryMonthly other = (empTimesalaryMonthly) that;
+        EmpTimesalaryMonthly other = (EmpTimesalaryMonthly) that;
         return (this.getId() == null ? other.getId() == null : this.getId().equals(other.getId()))
             && (this.getYearMonth() == null ? other.getYearMonth() == null : this.getYearMonth().equals(other.getYearMonth()))
             && (this.getEmpId() == null ? other.getEmpId() == null : this.getEmpId().equals(other.getEmpId()))
