+++ src/main/java/com/btjf/model/emp/EmpTimesalaryMonthlyExample.java
@@ -5,10 +5,10 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class empSubsidyMonthlyExample {
+public class EmpTimesalaryMonthlyExample {
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_emp_subsidy_monthly
+     * This field corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -16,7 +16,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_emp_subsidy_monthly
+     * This field corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -24,7 +24,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_emp_subsidy_monthly
+     * This field corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -32,17 +32,17 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
-    public empSubsidyMonthlyExample() {
+    public EmpTimesalaryMonthlyExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -52,7 +52,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -62,7 +62,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -72,7 +72,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -82,7 +82,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -92,7 +92,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -102,7 +102,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -114,7 +114,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -128,7 +128,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -139,7 +139,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_emp_subsidy_monthly
+     * This method corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -151,7 +151,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_emp_subsidy_monthly
+     * This class corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
@@ -656,63 +656,123 @@ public class empSubsidyMonthlyExample {
             return (Criteria) this;
         }
 
-        public Criteria andTypeIsNull() {
-            addCriterion("type is null");
+        public Criteria andNumIsNull() {
+            addCriterion("num is null");
             return (Criteria) this;
         }
 
-        public Criteria andTypeIsNotNull() {
-            addCriterion("type is not null");
+        public Criteria andNumIsNotNull() {
+            addCriterion("num is not null");
             return (Criteria) this;
         }
 
-        public Criteria andTypeEqualTo(Integer value) {
-            addCriterion("type =", value, "type");
+        public Criteria andNumEqualTo(BigDecimal value) {
+            addCriterion("num =", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeNotEqualTo(Integer value) {
-            addCriterion("type <>", value, "type");
+        public Criteria andNumNotEqualTo(BigDecimal value) {
+            addCriterion("num <>", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeGreaterThan(Integer value) {
-            addCriterion("type >", value, "type");
+        public Criteria andNumGreaterThan(BigDecimal value) {
+            addCriterion("num >", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeGreaterThanOrEqualTo(Integer value) {
-            addCriterion("type >=", value, "type");
+        public Criteria andNumGreaterThanOrEqualTo(BigDecimal value) {
+            addCriterion("num >=", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeLessThan(Integer value) {
-            addCriterion("type <", value, "type");
+        public Criteria andNumLessThan(BigDecimal value) {
+            addCriterion("num <", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeLessThanOrEqualTo(Integer value) {
-            addCriterion("type <=", value, "type");
+        public Criteria andNumLessThanOrEqualTo(BigDecimal value) {
+            addCriterion("num <=", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeIn(List<Integer> values) {
-            addCriterion("type in", values, "type");
+        public Criteria andNumIn(List<BigDecimal> values) {
+            addCriterion("num in", values, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeNotIn(List<Integer> values) {
-            addCriterion("type not in", values, "type");
+        public Criteria andNumNotIn(List<BigDecimal> values) {
+            addCriterion("num not in", values, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeBetween(Integer value1, Integer value2) {
-            addCriterion("type between", value1, value2, "type");
+        public Criteria andNumBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("num between", value1, value2, "num");
             return (Criteria) this;
         }
 
-        public Criteria andTypeNotBetween(Integer value1, Integer value2) {
-            addCriterion("type not between", value1, value2, "type");
+        public Criteria andNumNotBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("num not between", value1, value2, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceIsNull() {
+            addCriterion("price is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceIsNotNull() {
+            addCriterion("price is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceEqualTo(BigDecimal value) {
+            addCriterion("price =", value, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceNotEqualTo(BigDecimal value) {
+            addCriterion("price <>", value, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceGreaterThan(BigDecimal value) {
+            addCriterion("price >", value, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceGreaterThanOrEqualTo(BigDecimal value) {
+            addCriterion("price >=", value, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceLessThan(BigDecimal value) {
+            addCriterion("price <", value, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceLessThanOrEqualTo(BigDecimal value) {
+            addCriterion("price <=", value, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceIn(List<BigDecimal> values) {
+            addCriterion("price in", values, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceNotIn(List<BigDecimal> values) {
+            addCriterion("price not in", values, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("price between", value1, value2, "price");
+            return (Criteria) this;
+        }
+
+        public Criteria andPriceNotBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("price not between", value1, value2, "price");
             return (Criteria) this;
         }
 
@@ -846,73 +906,73 @@ public class empSubsidyMonthlyExample {
             return (Criteria) this;
         }
 
-        public Criteria andReasonIsNull() {
-            addCriterion("reason is null");
+        public Criteria andContentIsNull() {
+            addCriterion("content is null");
             return (Criteria) this;
         }
 
-        public Criteria andReasonIsNotNull() {
-            addCriterion("reason is not null");
+        public Criteria andContentIsNotNull() {
+            addCriterion("content is not null");
             return (Criteria) this;
         }
 
-        public Criteria andReasonEqualTo(String value) {
-            addCriterion("reason =", value, "reason");
+        public Criteria andContentEqualTo(String value) {
+            addCriterion("content =", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonNotEqualTo(String value) {
-            addCriterion("reason <>", value, "reason");
+        public Criteria andContentNotEqualTo(String value) {
+            addCriterion("content <>", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonGreaterThan(String value) {
-            addCriterion("reason >", value, "reason");
+        public Criteria andContentGreaterThan(String value) {
+            addCriterion("content >", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonGreaterThanOrEqualTo(String value) {
-            addCriterion("reason >=", value, "reason");
+        public Criteria andContentGreaterThanOrEqualTo(String value) {
+            addCriterion("content >=", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonLessThan(String value) {
-            addCriterion("reason <", value, "reason");
+        public Criteria andContentLessThan(String value) {
+            addCriterion("content <", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonLessThanOrEqualTo(String value) {
-            addCriterion("reason <=", value, "reason");
+        public Criteria andContentLessThanOrEqualTo(String value) {
+            addCriterion("content <=", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonLike(String value) {
-            addCriterion("reason like", value, "reason");
+        public Criteria andContentLike(String value) {
+            addCriterion("content like", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonNotLike(String value) {
-            addCriterion("reason not like", value, "reason");
+        public Criteria andContentNotLike(String value) {
+            addCriterion("content not like", value, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonIn(List<String> values) {
-            addCriterion("reason in", values, "reason");
+        public Criteria andContentIn(List<String> values) {
+            addCriterion("content in", values, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonNotIn(List<String> values) {
-            addCriterion("reason not in", values, "reason");
+        public Criteria andContentNotIn(List<String> values) {
+            addCriterion("content not in", values, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonBetween(String value1, String value2) {
-            addCriterion("reason between", value1, value2, "reason");
+        public Criteria andContentBetween(String value1, String value2) {
+            addCriterion("content between", value1, value2, "content");
             return (Criteria) this;
         }
 
-        public Criteria andReasonNotBetween(String value1, String value2) {
-            addCriterion("reason not between", value1, value2, "reason");
+        public Criteria andContentNotBetween(String value1, String value2) {
+            addCriterion("content not between", value1, value2, "content");
             return (Criteria) this;
         }
 
@@ -1381,8 +1441,8 @@ public class empSubsidyMonthlyExample {
             return (Criteria) this;
         }
 
-        public Criteria andReasonLikeInsensitive(String value) {
-            addCriterion("upper(reason) like", value.toUpperCase(), "reason");
+        public Criteria andContentLikeInsensitive(String value) {
+            addCriterion("upper(content) like", value.toUpperCase(), "content");
             return (Criteria) this;
         }
 
@@ -1399,7 +1459,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_emp_subsidy_monthly
+     * This class corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated do_not_delete_during_merge
      */
@@ -1412,7 +1472,7 @@ public class empSubsidyMonthlyExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_emp_subsidy_monthly
+     * This class corresponds to the database table t_emp_timesalary_monthly
      *
      * @mbg.generated
      */
