+++ src/main/java/com/btjf/model/emp/EmpWork.java
@@ -3,7 +3,7 @@ package com.btjf.model.emp;
 import java.io.Serializable;
 import java.util.Date;
 
-public class Empwork implements Serializable {
+public class EmpWork implements Serializable {
     /**
      *
      * This field was generated by MyBatis Generator.
@@ -194,7 +194,7 @@ public class Empwork implements Serializable {
         if (getClass() != that.getClass()) {
             return false;
         }
-        Empwork other = (Empwork) that;
+        EmpWork other = (EmpWork) that;
         return (this.getId() == null ? other.getId() == null : this.getId().equals(other.getId()))
             && (this.getName() == null ? other.getName() == null : this.getName().equals(other.getName()))
             && (this.getCreateTime() == null ? other.getCreateTime() == null : this.getCreateTime().equals(other.getCreateTime()))
