+++ src/main/java/com/btjf/model/emp/EmpWorkExample.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class EmpworkExample {
+public class EmpWorkExample {
     /**
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database table t_emp_work
@@ -35,7 +35,7 @@ public class EmpworkExample {
      *
      * @mbg.generated
      */
-    public EmpworkExample() {
+    public EmpWorkExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
