+++ src/main/java/com/btjf/model/pm/PmOutBill.java
@@ -3,7 +3,7 @@ package com.btjf.model.pm;
 import java.io.Serializable;
 import java.util.Date;
 
-public class pmOutBill implements Serializable {
+public class PmOutBill implements Serializable {
     /**
      *
      * This field was generated by MyBatis Generator.
@@ -557,7 +557,7 @@ public class pmOutBill implements Serializable {
         if (getClass() != that.getClass()) {
             return false;
         }
-        pmOutBill other = (pmOutBill) that;
+        PmOutBill other = (PmOutBill) that;
         return (this.getId() == null ? other.getId() == null : this.getId().equals(other.getId()))
             && (this.getOrderId() == null ? other.getOrderId() == null : this.getOrderId().equals(other.getOrderId()))
             && (this.getOrderNo() == null ? other.getOrderNo() == null : this.getOrderNo().equals(other.getOrderNo()))
