+++ src/main/java/com/btjf/model/pm/PmOutBillDetail.java
@@ -3,7 +3,7 @@ package com.btjf.model.pm;
 import java.io.Serializable;
 import java.util.Date;
 
-public class PmoutBillDetail implements Serializable {
+public class PmOutBillDetail implements Serializable {
     /**
      *
      * This field was generated by MyBatis Generator.
@@ -623,7 +623,7 @@ public class PmoutBillDetail implements Serializable {
         if (getClass() != that.getClass()) {
             return false;
         }
-        PmoutBillDetail other = (PmoutBillDetail) that;
+        PmOutBillDetail other = (PmOutBillDetail) that;
         return (this.getId() == null ? other.getId() == null : this.getId().equals(other.getId()))
             && (this.getBillId() == null ? other.getBillId() == null : this.getBillId().equals(other.getBillId()))
             && (this.getPmId() == null ? other.getPmId() == null : this.getPmId().equals(other.getPmId()))
