+++ src/main/java/com/btjf/model/pm/PmOutBillDetailExample.java
@@ -1,14 +1,13 @@
-package com.btjf.model.product;
+package com.btjf.model.pm;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class ProductpmExample {
+public class PmOutBillDetailExample {
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_productpm
+     * This field corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -16,7 +15,7 @@ public class ProductpmExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_productpm
+     * This field corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -24,7 +23,7 @@ public class ProductpmExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_productpm
+     * This field corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -32,17 +31,17 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
-    public ProductpmExample() {
+    public PmOutBillDetailExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -52,7 +51,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -62,7 +61,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -72,7 +71,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -82,7 +81,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -92,7 +91,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -102,7 +101,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -114,7 +113,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -128,7 +127,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -139,7 +138,7 @@ public class ProductpmExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_productpm
+     * This method corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -151,7 +150,7 @@ public class ProductpmExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_productpm
+     * This class corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
@@ -256,133 +255,63 @@ public class ProductpmExample {
             return (Criteria) this;
         }
 
-        public Criteria andProductIdIsNull() {
-            addCriterion("productId is null");
+        public Criteria andBillIdIsNull() {
+            addCriterion("billId is null");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdIsNotNull() {
-            addCriterion("productId is not null");
+        public Criteria andBillIdIsNotNull() {
+            addCriterion("billId is not null");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdEqualTo(Integer value) {
-            addCriterion("productId =", value, "productId");
+        public Criteria andBillIdEqualTo(Integer value) {
+            addCriterion("billId =", value, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdNotEqualTo(Integer value) {
-            addCriterion("productId <>", value, "productId");
+        public Criteria andBillIdNotEqualTo(Integer value) {
+            addCriterion("billId <>", value, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdGreaterThan(Integer value) {
-            addCriterion("productId >", value, "productId");
+        public Criteria andBillIdGreaterThan(Integer value) {
+            addCriterion("billId >", value, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdGreaterThanOrEqualTo(Integer value) {
-            addCriterion("productId >=", value, "productId");
+        public Criteria andBillIdGreaterThanOrEqualTo(Integer value) {
+            addCriterion("billId >=", value, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdLessThan(Integer value) {
-            addCriterion("productId <", value, "productId");
+        public Criteria andBillIdLessThan(Integer value) {
+            addCriterion("billId <", value, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdLessThanOrEqualTo(Integer value) {
-            addCriterion("productId <=", value, "productId");
+        public Criteria andBillIdLessThanOrEqualTo(Integer value) {
+            addCriterion("billId <=", value, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdIn(List<Integer> values) {
-            addCriterion("productId in", values, "productId");
+        public Criteria andBillIdIn(List<Integer> values) {
+            addCriterion("billId in", values, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdNotIn(List<Integer> values) {
-            addCriterion("productId not in", values, "productId");
+        public Criteria andBillIdNotIn(List<Integer> values) {
+            addCriterion("billId not in", values, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdBetween(Integer value1, Integer value2) {
-            addCriterion("productId between", value1, value2, "productId");
+        public Criteria andBillIdBetween(Integer value1, Integer value2) {
+            addCriterion("billId between", value1, value2, "billId");
             return (Criteria) this;
         }
 
-        public Criteria andProductIdNotBetween(Integer value1, Integer value2) {
-            addCriterion("productId not between", value1, value2, "productId");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoIsNull() {
-            addCriterion("productNo is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoIsNotNull() {
-            addCriterion("productNo is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoEqualTo(String value) {
-            addCriterion("productNo =", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoNotEqualTo(String value) {
-            addCriterion("productNo <>", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoGreaterThan(String value) {
-            addCriterion("productNo >", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoGreaterThanOrEqualTo(String value) {
-            addCriterion("productNo >=", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoLessThan(String value) {
-            addCriterion("productNo <", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoLessThanOrEqualTo(String value) {
-            addCriterion("productNo <=", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoLike(String value) {
-            addCriterion("productNo like", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoNotLike(String value) {
-            addCriterion("productNo not like", value, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoIn(List<String> values) {
-            addCriterion("productNo in", values, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoNotIn(List<String> values) {
-            addCriterion("productNo not in", values, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoBetween(String value1, String value2) {
-            addCriterion("productNo between", value1, value2, "productNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andProductNoNotBetween(String value1, String value2) {
-            addCriterion("productNo not between", value1, value2, "productNo");
+        public Criteria andBillIdNotBetween(Integer value1, Integer value2) {
+            addCriterion("billId not between", value1, value2, "billId");
             return (Criteria) this;
         }
 
@@ -586,73 +515,253 @@ public class ProductpmExample {
             return (Criteria) this;
         }
 
-        public Criteria andTypeIsNull() {
-            addCriterion("type is null");
+        public Criteria andPmBatchNoIsNull() {
+            addCriterion("pmBatchNo is null");
             return (Criteria) this;
         }
 
-        public Criteria andTypeIsNotNull() {
-            addCriterion("type is not null");
+        public Criteria andPmBatchNoIsNotNull() {
+            addCriterion("pmBatchNo is not null");
             return (Criteria) this;
         }
 
-        public Criteria andTypeEqualTo(String value) {
-            addCriterion("type =", value, "type");
+        public Criteria andPmBatchNoEqualTo(String value) {
+            addCriterion("pmBatchNo =", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeNotEqualTo(String value) {
-            addCriterion("type <>", value, "type");
+        public Criteria andPmBatchNoNotEqualTo(String value) {
+            addCriterion("pmBatchNo <>", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeGreaterThan(String value) {
-            addCriterion("type >", value, "type");
+        public Criteria andPmBatchNoGreaterThan(String value) {
+            addCriterion("pmBatchNo >", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeGreaterThanOrEqualTo(String value) {
-            addCriterion("type >=", value, "type");
+        public Criteria andPmBatchNoGreaterThanOrEqualTo(String value) {
+            addCriterion("pmBatchNo >=", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeLessThan(String value) {
-            addCriterion("type <", value, "type");
+        public Criteria andPmBatchNoLessThan(String value) {
+            addCriterion("pmBatchNo <", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeLessThanOrEqualTo(String value) {
-            addCriterion("type <=", value, "type");
+        public Criteria andPmBatchNoLessThanOrEqualTo(String value) {
+            addCriterion("pmBatchNo <=", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeLike(String value) {
-            addCriterion("type like", value, "type");
+        public Criteria andPmBatchNoLike(String value) {
+            addCriterion("pmBatchNo like", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeNotLike(String value) {
-            addCriterion("type not like", value, "type");
+        public Criteria andPmBatchNoNotLike(String value) {
+            addCriterion("pmBatchNo not like", value, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeIn(List<String> values) {
-            addCriterion("type in", values, "type");
+        public Criteria andPmBatchNoIn(List<String> values) {
+            addCriterion("pmBatchNo in", values, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeNotIn(List<String> values) {
-            addCriterion("type not in", values, "type");
+        public Criteria andPmBatchNoNotIn(List<String> values) {
+            addCriterion("pmBatchNo not in", values, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeBetween(String value1, String value2) {
-            addCriterion("type between", value1, value2, "type");
+        public Criteria andPmBatchNoBetween(String value1, String value2) {
+            addCriterion("pmBatchNo between", value1, value2, "pmBatchNo");
             return (Criteria) this;
         }
 
-        public Criteria andTypeNotBetween(String value1, String value2) {
-            addCriterion("type not between", value1, value2, "type");
+        public Criteria andPmBatchNoNotBetween(String value1, String value2) {
+            addCriterion("pmBatchNo not between", value1, value2, "pmBatchNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumIsNull() {
+            addCriterion("num is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumIsNotNull() {
+            addCriterion("num is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumEqualTo(Integer value) {
+            addCriterion("num =", value, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumNotEqualTo(Integer value) {
+            addCriterion("num <>", value, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumGreaterThan(Integer value) {
+            addCriterion("num >", value, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumGreaterThanOrEqualTo(Integer value) {
+            addCriterion("num >=", value, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumLessThan(Integer value) {
+            addCriterion("num <", value, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumLessThanOrEqualTo(Integer value) {
+            addCriterion("num <=", value, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumIn(List<Integer> values) {
+            addCriterion("num in", values, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumNotIn(List<Integer> values) {
+            addCriterion("num not in", values, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumBetween(Integer value1, Integer value2) {
+            addCriterion("num between", value1, value2, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andNumNotBetween(Integer value1, Integer value2) {
+            addCriterion("num not between", value1, value2, "num");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumIsNull() {
+            addCriterion("perNum is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumIsNotNull() {
+            addCriterion("perNum is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumEqualTo(Integer value) {
+            addCriterion("perNum =", value, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumNotEqualTo(Integer value) {
+            addCriterion("perNum <>", value, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumGreaterThan(Integer value) {
+            addCriterion("perNum >", value, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumGreaterThanOrEqualTo(Integer value) {
+            addCriterion("perNum >=", value, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumLessThan(Integer value) {
+            addCriterion("perNum <", value, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumLessThanOrEqualTo(Integer value) {
+            addCriterion("perNum <=", value, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumIn(List<Integer> values) {
+            addCriterion("perNum in", values, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumNotIn(List<Integer> values) {
+            addCriterion("perNum not in", values, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumBetween(Integer value1, Integer value2) {
+            addCriterion("perNum between", value1, value2, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andPerNumNotBetween(Integer value1, Integer value2) {
+            addCriterion("perNum not between", value1, value2, "perNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumIsNull() {
+            addCriterion("backNum is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumIsNotNull() {
+            addCriterion("backNum is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumEqualTo(Integer value) {
+            addCriterion("backNum =", value, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumNotEqualTo(Integer value) {
+            addCriterion("backNum <>", value, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumGreaterThan(Integer value) {
+            addCriterion("backNum >", value, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumGreaterThanOrEqualTo(Integer value) {
+            addCriterion("backNum >=", value, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumLessThan(Integer value) {
+            addCriterion("backNum <", value, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumLessThanOrEqualTo(Integer value) {
+            addCriterion("backNum <=", value, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumIn(List<Integer> values) {
+            addCriterion("backNum in", values, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumNotIn(List<Integer> values) {
+            addCriterion("backNum not in", values, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumBetween(Integer value1, Integer value2) {
+            addCriterion("backNum between", value1, value2, "backNum");
+            return (Criteria) this;
+        }
+
+        public Criteria andBackNumNotBetween(Integer value1, Integer value2) {
+            addCriterion("backNum not between", value1, value2, "backNum");
             return (Criteria) this;
         }
 
@@ -796,183 +905,213 @@ public class ProductpmExample {
             return (Criteria) this;
         }
 
-        public Criteria andStatusIsNull() {
-            addCriterion("status is null");
+        public Criteria andOutInfoIsNull() {
+            addCriterion("outInfo is null");
             return (Criteria) this;
         }
 
-        public Criteria andStatusIsNotNull() {
-            addCriterion("status is not null");
+        public Criteria andOutInfoIsNotNull() {
+            addCriterion("outInfo is not null");
             return (Criteria) this;
         }
 
-        public Criteria andStatusEqualTo(Integer value) {
-            addCriterion("status =", value, "status");
+        public Criteria andOutInfoEqualTo(String value) {
+            addCriterion("outInfo =", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusNotEqualTo(Integer value) {
-            addCriterion("status <>", value, "status");
+        public Criteria andOutInfoNotEqualTo(String value) {
+            addCriterion("outInfo <>", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusGreaterThan(Integer value) {
-            addCriterion("status >", value, "status");
+        public Criteria andOutInfoGreaterThan(String value) {
+            addCriterion("outInfo >", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusGreaterThanOrEqualTo(Integer value) {
-            addCriterion("status >=", value, "status");
+        public Criteria andOutInfoGreaterThanOrEqualTo(String value) {
+            addCriterion("outInfo >=", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusLessThan(Integer value) {
-            addCriterion("status <", value, "status");
+        public Criteria andOutInfoLessThan(String value) {
+            addCriterion("outInfo <", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusLessThanOrEqualTo(Integer value) {
-            addCriterion("status <=", value, "status");
+        public Criteria andOutInfoLessThanOrEqualTo(String value) {
+            addCriterion("outInfo <=", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusIn(List<Integer> values) {
-            addCriterion("status in", values, "status");
+        public Criteria andOutInfoLike(String value) {
+            addCriterion("outInfo like", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusNotIn(List<Integer> values) {
-            addCriterion("status not in", values, "status");
+        public Criteria andOutInfoNotLike(String value) {
+            addCriterion("outInfo not like", value, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusBetween(Integer value1, Integer value2) {
-            addCriterion("status between", value1, value2, "status");
+        public Criteria andOutInfoIn(List<String> values) {
+            addCriterion("outInfo in", values, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andStatusNotBetween(Integer value1, Integer value2) {
-            addCriterion("status not between", value1, value2, "status");
+        public Criteria andOutInfoNotIn(List<String> values) {
+            addCriterion("outInfo not in", values, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andNumIsNull() {
-            addCriterion("num is null");
+        public Criteria andOutInfoBetween(String value1, String value2) {
+            addCriterion("outInfo between", value1, value2, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andNumIsNotNull() {
-            addCriterion("num is not null");
+        public Criteria andOutInfoNotBetween(String value1, String value2) {
+            addCriterion("outInfo not between", value1, value2, "outInfo");
             return (Criteria) this;
         }
 
-        public Criteria andNumEqualTo(BigDecimal value) {
-            addCriterion("num =", value, "num");
+        public Criteria andReceiverIsNull() {
+            addCriterion("receiver is null");
             return (Criteria) this;
         }
 
-        public Criteria andNumNotEqualTo(BigDecimal value) {
-            addCriterion("num <>", value, "num");
+        public Criteria andReceiverIsNotNull() {
+            addCriterion("receiver is not null");
             return (Criteria) this;
         }
 
-        public Criteria andNumGreaterThan(BigDecimal value) {
-            addCriterion("num >", value, "num");
+        public Criteria andReceiverEqualTo(String value) {
+            addCriterion("receiver =", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andNumGreaterThanOrEqualTo(BigDecimal value) {
-            addCriterion("num >=", value, "num");
+        public Criteria andReceiverNotEqualTo(String value) {
+            addCriterion("receiver <>", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andNumLessThan(BigDecimal value) {
-            addCriterion("num <", value, "num");
+        public Criteria andReceiverGreaterThan(String value) {
+            addCriterion("receiver >", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andNumLessThanOrEqualTo(BigDecimal value) {
-            addCriterion("num <=", value, "num");
+        public Criteria andReceiverGreaterThanOrEqualTo(String value) {
+            addCriterion("receiver >=", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andNumIn(List<BigDecimal> values) {
-            addCriterion("num in", values, "num");
+        public Criteria andReceiverLessThan(String value) {
+            addCriterion("receiver <", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andNumNotIn(List<BigDecimal> values) {
-            addCriterion("num not in", values, "num");
+        public Criteria andReceiverLessThanOrEqualTo(String value) {
+            addCriterion("receiver <=", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andNumBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("num between", value1, value2, "num");
+        public Criteria andReceiverLike(String value) {
+            addCriterion("receiver like", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andNumNotBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("num not between", value1, value2, "num");
+        public Criteria andReceiverNotLike(String value) {
+            addCriterion("receiver not like", value, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumIsNull() {
-            addCriterion("unitNum is null");
+        public Criteria andReceiverIn(List<String> values) {
+            addCriterion("receiver in", values, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumIsNotNull() {
-            addCriterion("unitNum is not null");
+        public Criteria andReceiverNotIn(List<String> values) {
+            addCriterion("receiver not in", values, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumEqualTo(BigDecimal value) {
-            addCriterion("unitNum =", value, "unitNum");
+        public Criteria andReceiverBetween(String value1, String value2) {
+            addCriterion("receiver between", value1, value2, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumNotEqualTo(BigDecimal value) {
-            addCriterion("unitNum <>", value, "unitNum");
+        public Criteria andReceiverNotBetween(String value1, String value2) {
+            addCriterion("receiver not between", value1, value2, "receiver");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumGreaterThan(BigDecimal value) {
-            addCriterion("unitNum >", value, "unitNum");
+        public Criteria andSenderIsNull() {
+            addCriterion("sender is null");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumGreaterThanOrEqualTo(BigDecimal value) {
-            addCriterion("unitNum >=", value, "unitNum");
+        public Criteria andSenderIsNotNull() {
+            addCriterion("sender is not null");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumLessThan(BigDecimal value) {
-            addCriterion("unitNum <", value, "unitNum");
+        public Criteria andSenderEqualTo(String value) {
+            addCriterion("sender =", value, "sender");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumLessThanOrEqualTo(BigDecimal value) {
-            addCriterion("unitNum <=", value, "unitNum");
+        public Criteria andSenderNotEqualTo(String value) {
+            addCriterion("sender <>", value, "sender");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumIn(List<BigDecimal> values) {
-            addCriterion("unitNum in", values, "unitNum");
+        public Criteria andSenderGreaterThan(String value) {
+            addCriterion("sender >", value, "sender");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumNotIn(List<BigDecimal> values) {
-            addCriterion("unitNum not in", values, "unitNum");
+        public Criteria andSenderGreaterThanOrEqualTo(String value) {
+            addCriterion("sender >=", value, "sender");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("unitNum between", value1, value2, "unitNum");
+        public Criteria andSenderLessThan(String value) {
+            addCriterion("sender <", value, "sender");
             return (Criteria) this;
         }
 
-        public Criteria andUnitNumNotBetween(BigDecimal value1, BigDecimal value2) {
-            addCriterion("unitNum not between", value1, value2, "unitNum");
+        public Criteria andSenderLessThanOrEqualTo(String value) {
+            addCriterion("sender <=", value, "sender");
+            return (Criteria) this;
+        }
+
+        public Criteria andSenderLike(String value) {
+            addCriterion("sender like", value, "sender");
+            return (Criteria) this;
+        }
+
+        public Criteria andSenderNotLike(String value) {
+            addCriterion("sender not like", value, "sender");
+            return (Criteria) this;
+        }
+
+        public Criteria andSenderIn(List<String> values) {
+            addCriterion("sender in", values, "sender");
+            return (Criteria) this;
+        }
+
+        public Criteria andSenderNotIn(List<String> values) {
+            addCriterion("sender not in", values, "sender");
+            return (Criteria) this;
+        }
+
+        public Criteria andSenderBetween(String value1, String value2) {
+            addCriterion("sender between", value1, value2, "sender");
+            return (Criteria) this;
+        }
+
+        public Criteria andSenderNotBetween(String value1, String value2) {
+            addCriterion("sender not between", value1, value2, "sender");
             return (Criteria) this;
         }
 
@@ -1226,11 +1365,6 @@ public class ProductpmExample {
             return (Criteria) this;
         }
 
-        public Criteria andProductNoLikeInsensitive(String value) {
-            addCriterion("upper(productNo) like", value.toUpperCase(), "productNo");
-            return (Criteria) this;
-        }
-
         public Criteria andPmNoLikeInsensitive(String value) {
             addCriterion("upper(pmNo) like", value.toUpperCase(), "pmNo");
             return (Criteria) this;
@@ -1241,8 +1375,8 @@ public class ProductpmExample {
             return (Criteria) this;
         }
 
-        public Criteria andTypeLikeInsensitive(String value) {
-            addCriterion("upper(type) like", value.toUpperCase(), "type");
+        public Criteria andPmBatchNoLikeInsensitive(String value) {
+            addCriterion("upper(pmBatchNo) like", value.toUpperCase(), "pmBatchNo");
             return (Criteria) this;
         }
 
@@ -1256,6 +1390,21 @@ public class ProductpmExample {
             return (Criteria) this;
         }
 
+        public Criteria andOutInfoLikeInsensitive(String value) {
+            addCriterion("upper(outInfo) like", value.toUpperCase(), "outInfo");
+            return (Criteria) this;
+        }
+
+        public Criteria andReceiverLikeInsensitive(String value) {
+            addCriterion("upper(receiver) like", value.toUpperCase(), "receiver");
+            return (Criteria) this;
+        }
+
+        public Criteria andSenderLikeInsensitive(String value) {
+            addCriterion("upper(sender) like", value.toUpperCase(), "sender");
+            return (Criteria) this;
+        }
+
         public Criteria andOperatorLikeInsensitive(String value) {
             addCriterion("upper(operator) like", value.toUpperCase(), "operator");
             return (Criteria) this;
@@ -1264,7 +1413,7 @@ public class ProductpmExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_productpm
+     * This class corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated do_not_delete_during_merge
      */
@@ -1277,7 +1426,7 @@ public class ProductpmExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_productpm
+     * This class corresponds to the database table t_pmout_bill_detail
      *
      * @mbg.generated
      */
