+++ src/main/java/com/btjf/model/pm/PmOutBillExample.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class pmOutBillExample {
+public class PmOutBillExample {
     /**
      * This field was generated by MyBatis Generator.
      * This field corresponds to the database table t_pmout_bill
@@ -35,7 +35,7 @@ public class pmOutBillExample {
      *
      * @mbg.generated
      */
-    public pmOutBillExample() {
+    public PmOutBillExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
