+++ src/main/java/com/btjf/model/product/ProductPm.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.Date;
 
-public class Productpm implements Serializable {
+public class ProductPm implements Serializable {
     /**
      *
      * This field was generated by MyBatis Generator.
@@ -558,7 +558,7 @@ public class Productpm implements Serializable {
         if (getClass() != that.getClass()) {
             return false;
         }
-        Productpm other = (Productpm) that;
+        ProductPm other = (ProductPm) that;
         return (this.getId() == null ? other.getId() == null : this.getId().equals(other.getId()))
             && (this.getProductId() == null ? other.getProductId() == null : this.getProductId().equals(other.getProductId()))
             && (this.getProductNo() == null ? other.getProductNo() == null : this.getProductNo().equals(other.getProductNo()))
