+++ src/main/java/com/btjf/model/product/ProductPmExample.java
@@ -1,13 +1,14 @@
-package com.btjf.model.pm;
+package com.btjf.model.product;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class PmoutBillDetailExample {
+public class ProductPmExample {
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_pmout_bill_detail
+     * This field corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -15,7 +16,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_pmout_bill_detail
+     * This field corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -23,7 +24,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_pmout_bill_detail
+     * This field corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -31,17 +32,17 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
-    public PmoutBillDetailExample() {
+    public ProductPmExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -51,7 +52,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -61,7 +62,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -71,7 +72,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -81,7 +82,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -91,7 +92,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -101,7 +102,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -113,7 +114,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -127,7 +128,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -138,7 +139,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_pmout_bill_detail
+     * This method corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -150,7 +151,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_pmout_bill_detail
+     * This class corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
@@ -255,63 +256,133 @@ public class PmoutBillDetailExample {
             return (Criteria) this;
         }
 
-        public Criteria andBillIdIsNull() {
-            addCriterion("billId is null");
+        public Criteria andProductIdIsNull() {
+            addCriterion("productId is null");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdIsNotNull() {
-            addCriterion("billId is not null");
+        public Criteria andProductIdIsNotNull() {
+            addCriterion("productId is not null");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdEqualTo(Integer value) {
-            addCriterion("billId =", value, "billId");
+        public Criteria andProductIdEqualTo(Integer value) {
+            addCriterion("productId =", value, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdNotEqualTo(Integer value) {
-            addCriterion("billId <>", value, "billId");
+        public Criteria andProductIdNotEqualTo(Integer value) {
+            addCriterion("productId <>", value, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdGreaterThan(Integer value) {
-            addCriterion("billId >", value, "billId");
+        public Criteria andProductIdGreaterThan(Integer value) {
+            addCriterion("productId >", value, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdGreaterThanOrEqualTo(Integer value) {
-            addCriterion("billId >=", value, "billId");
+        public Criteria andProductIdGreaterThanOrEqualTo(Integer value) {
+            addCriterion("productId >=", value, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdLessThan(Integer value) {
-            addCriterion("billId <", value, "billId");
+        public Criteria andProductIdLessThan(Integer value) {
+            addCriterion("productId <", value, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdLessThanOrEqualTo(Integer value) {
-            addCriterion("billId <=", value, "billId");
+        public Criteria andProductIdLessThanOrEqualTo(Integer value) {
+            addCriterion("productId <=", value, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdIn(List<Integer> values) {
-            addCriterion("billId in", values, "billId");
+        public Criteria andProductIdIn(List<Integer> values) {
+            addCriterion("productId in", values, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdNotIn(List<Integer> values) {
-            addCriterion("billId not in", values, "billId");
+        public Criteria andProductIdNotIn(List<Integer> values) {
+            addCriterion("productId not in", values, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdBetween(Integer value1, Integer value2) {
-            addCriterion("billId between", value1, value2, "billId");
+        public Criteria andProductIdBetween(Integer value1, Integer value2) {
+            addCriterion("productId between", value1, value2, "productId");
             return (Criteria) this;
         }
 
-        public Criteria andBillIdNotBetween(Integer value1, Integer value2) {
-            addCriterion("billId not between", value1, value2, "billId");
+        public Criteria andProductIdNotBetween(Integer value1, Integer value2) {
+            addCriterion("productId not between", value1, value2, "productId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoIsNull() {
+            addCriterion("productNo is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoIsNotNull() {
+            addCriterion("productNo is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoEqualTo(String value) {
+            addCriterion("productNo =", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoNotEqualTo(String value) {
+            addCriterion("productNo <>", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoGreaterThan(String value) {
+            addCriterion("productNo >", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoGreaterThanOrEqualTo(String value) {
+            addCriterion("productNo >=", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoLessThan(String value) {
+            addCriterion("productNo <", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoLessThanOrEqualTo(String value) {
+            addCriterion("productNo <=", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoLike(String value) {
+            addCriterion("productNo like", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoNotLike(String value) {
+            addCriterion("productNo not like", value, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoIn(List<String> values) {
+            addCriterion("productNo in", values, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoNotIn(List<String> values) {
+            addCriterion("productNo not in", values, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoBetween(String value1, String value2) {
+            addCriterion("productNo between", value1, value2, "productNo");
+            return (Criteria) this;
+        }
+
+        public Criteria andProductNoNotBetween(String value1, String value2) {
+            addCriterion("productNo not between", value1, value2, "productNo");
             return (Criteria) this;
         }
 
@@ -515,253 +586,73 @@ public class PmoutBillDetailExample {
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoIsNull() {
-            addCriterion("pmBatchNo is null");
+        public Criteria andTypeIsNull() {
+            addCriterion("type is null");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoIsNotNull() {
-            addCriterion("pmBatchNo is not null");
+        public Criteria andTypeIsNotNull() {
+            addCriterion("type is not null");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoEqualTo(String value) {
-            addCriterion("pmBatchNo =", value, "pmBatchNo");
+        public Criteria andTypeEqualTo(String value) {
+            addCriterion("type =", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoNotEqualTo(String value) {
-            addCriterion("pmBatchNo <>", value, "pmBatchNo");
+        public Criteria andTypeNotEqualTo(String value) {
+            addCriterion("type <>", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoGreaterThan(String value) {
-            addCriterion("pmBatchNo >", value, "pmBatchNo");
+        public Criteria andTypeGreaterThan(String value) {
+            addCriterion("type >", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoGreaterThanOrEqualTo(String value) {
-            addCriterion("pmBatchNo >=", value, "pmBatchNo");
+        public Criteria andTypeGreaterThanOrEqualTo(String value) {
+            addCriterion("type >=", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoLessThan(String value) {
-            addCriterion("pmBatchNo <", value, "pmBatchNo");
+        public Criteria andTypeLessThan(String value) {
+            addCriterion("type <", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoLessThanOrEqualTo(String value) {
-            addCriterion("pmBatchNo <=", value, "pmBatchNo");
+        public Criteria andTypeLessThanOrEqualTo(String value) {
+            addCriterion("type <=", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoLike(String value) {
-            addCriterion("pmBatchNo like", value, "pmBatchNo");
+        public Criteria andTypeLike(String value) {
+            addCriterion("type like", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoNotLike(String value) {
-            addCriterion("pmBatchNo not like", value, "pmBatchNo");
+        public Criteria andTypeNotLike(String value) {
+            addCriterion("type not like", value, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoIn(List<String> values) {
-            addCriterion("pmBatchNo in", values, "pmBatchNo");
+        public Criteria andTypeIn(List<String> values) {
+            addCriterion("type in", values, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoNotIn(List<String> values) {
-            addCriterion("pmBatchNo not in", values, "pmBatchNo");
+        public Criteria andTypeNotIn(List<String> values) {
+            addCriterion("type not in", values, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoBetween(String value1, String value2) {
-            addCriterion("pmBatchNo between", value1, value2, "pmBatchNo");
+        public Criteria andTypeBetween(String value1, String value2) {
+            addCriterion("type between", value1, value2, "type");
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoNotBetween(String value1, String value2) {
-            addCriterion("pmBatchNo not between", value1, value2, "pmBatchNo");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumIsNull() {
-            addCriterion("num is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumIsNotNull() {
-            addCriterion("num is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumEqualTo(Integer value) {
-            addCriterion("num =", value, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumNotEqualTo(Integer value) {
-            addCriterion("num <>", value, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumGreaterThan(Integer value) {
-            addCriterion("num >", value, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumGreaterThanOrEqualTo(Integer value) {
-            addCriterion("num >=", value, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumLessThan(Integer value) {
-            addCriterion("num <", value, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumLessThanOrEqualTo(Integer value) {
-            addCriterion("num <=", value, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumIn(List<Integer> values) {
-            addCriterion("num in", values, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumNotIn(List<Integer> values) {
-            addCriterion("num not in", values, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumBetween(Integer value1, Integer value2) {
-            addCriterion("num between", value1, value2, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andNumNotBetween(Integer value1, Integer value2) {
-            addCriterion("num not between", value1, value2, "num");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumIsNull() {
-            addCriterion("perNum is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumIsNotNull() {
-            addCriterion("perNum is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumEqualTo(Integer value) {
-            addCriterion("perNum =", value, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumNotEqualTo(Integer value) {
-            addCriterion("perNum <>", value, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumGreaterThan(Integer value) {
-            addCriterion("perNum >", value, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumGreaterThanOrEqualTo(Integer value) {
-            addCriterion("perNum >=", value, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumLessThan(Integer value) {
-            addCriterion("perNum <", value, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumLessThanOrEqualTo(Integer value) {
-            addCriterion("perNum <=", value, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumIn(List<Integer> values) {
-            addCriterion("perNum in", values, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumNotIn(List<Integer> values) {
-            addCriterion("perNum not in", values, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumBetween(Integer value1, Integer value2) {
-            addCriterion("perNum between", value1, value2, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andPerNumNotBetween(Integer value1, Integer value2) {
-            addCriterion("perNum not between", value1, value2, "perNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumIsNull() {
-            addCriterion("backNum is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumIsNotNull() {
-            addCriterion("backNum is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumEqualTo(Integer value) {
-            addCriterion("backNum =", value, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumNotEqualTo(Integer value) {
-            addCriterion("backNum <>", value, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumGreaterThan(Integer value) {
-            addCriterion("backNum >", value, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumGreaterThanOrEqualTo(Integer value) {
-            addCriterion("backNum >=", value, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumLessThan(Integer value) {
-            addCriterion("backNum <", value, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumLessThanOrEqualTo(Integer value) {
-            addCriterion("backNum <=", value, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumIn(List<Integer> values) {
-            addCriterion("backNum in", values, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumNotIn(List<Integer> values) {
-            addCriterion("backNum not in", values, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumBetween(Integer value1, Integer value2) {
-            addCriterion("backNum between", value1, value2, "backNum");
-            return (Criteria) this;
-        }
-
-        public Criteria andBackNumNotBetween(Integer value1, Integer value2) {
-            addCriterion("backNum not between", value1, value2, "backNum");
+        public Criteria andTypeNotBetween(String value1, String value2) {
+            addCriterion("type not between", value1, value2, "type");
             return (Criteria) this;
         }
 
@@ -905,213 +796,183 @@ public class PmoutBillDetailExample {
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoIsNull() {
-            addCriterion("outInfo is null");
+        public Criteria andStatusIsNull() {
+            addCriterion("status is null");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoIsNotNull() {
-            addCriterion("outInfo is not null");
+        public Criteria andStatusIsNotNull() {
+            addCriterion("status is not null");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoEqualTo(String value) {
-            addCriterion("outInfo =", value, "outInfo");
+        public Criteria andStatusEqualTo(Integer value) {
+            addCriterion("status =", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoNotEqualTo(String value) {
-            addCriterion("outInfo <>", value, "outInfo");
+        public Criteria andStatusNotEqualTo(Integer value) {
+            addCriterion("status <>", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoGreaterThan(String value) {
-            addCriterion("outInfo >", value, "outInfo");
+        public Criteria andStatusGreaterThan(Integer value) {
+            addCriterion("status >", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoGreaterThanOrEqualTo(String value) {
-            addCriterion("outInfo >=", value, "outInfo");
+        public Criteria andStatusGreaterThanOrEqualTo(Integer value) {
+            addCriterion("status >=", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoLessThan(String value) {
-            addCriterion("outInfo <", value, "outInfo");
+        public Criteria andStatusLessThan(Integer value) {
+            addCriterion("status <", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoLessThanOrEqualTo(String value) {
-            addCriterion("outInfo <=", value, "outInfo");
+        public Criteria andStatusLessThanOrEqualTo(Integer value) {
+            addCriterion("status <=", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoLike(String value) {
-            addCriterion("outInfo like", value, "outInfo");
+        public Criteria andStatusIn(List<Integer> values) {
+            addCriterion("status in", values, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoNotLike(String value) {
-            addCriterion("outInfo not like", value, "outInfo");
+        public Criteria andStatusNotIn(List<Integer> values) {
+            addCriterion("status not in", values, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoIn(List<String> values) {
-            addCriterion("outInfo in", values, "outInfo");
+        public Criteria andStatusBetween(Integer value1, Integer value2) {
+            addCriterion("status between", value1, value2, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoNotIn(List<String> values) {
-            addCriterion("outInfo not in", values, "outInfo");
+        public Criteria andStatusNotBetween(Integer value1, Integer value2) {
+            addCriterion("status not between", value1, value2, "status");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoBetween(String value1, String value2) {
-            addCriterion("outInfo between", value1, value2, "outInfo");
+        public Criteria andNumIsNull() {
+            addCriterion("num is null");
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoNotBetween(String value1, String value2) {
-            addCriterion("outInfo not between", value1, value2, "outInfo");
+        public Criteria andNumIsNotNull() {
+            addCriterion("num is not null");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverIsNull() {
-            addCriterion("receiver is null");
+        public Criteria andNumEqualTo(BigDecimal value) {
+            addCriterion("num =", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverIsNotNull() {
-            addCriterion("receiver is not null");
+        public Criteria andNumNotEqualTo(BigDecimal value) {
+            addCriterion("num <>", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverEqualTo(String value) {
-            addCriterion("receiver =", value, "receiver");
+        public Criteria andNumGreaterThan(BigDecimal value) {
+            addCriterion("num >", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverNotEqualTo(String value) {
-            addCriterion("receiver <>", value, "receiver");
+        public Criteria andNumGreaterThanOrEqualTo(BigDecimal value) {
+            addCriterion("num >=", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverGreaterThan(String value) {
-            addCriterion("receiver >", value, "receiver");
+        public Criteria andNumLessThan(BigDecimal value) {
+            addCriterion("num <", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverGreaterThanOrEqualTo(String value) {
-            addCriterion("receiver >=", value, "receiver");
+        public Criteria andNumLessThanOrEqualTo(BigDecimal value) {
+            addCriterion("num <=", value, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverLessThan(String value) {
-            addCriterion("receiver <", value, "receiver");
+        public Criteria andNumIn(List<BigDecimal> values) {
+            addCriterion("num in", values, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverLessThanOrEqualTo(String value) {
-            addCriterion("receiver <=", value, "receiver");
+        public Criteria andNumNotIn(List<BigDecimal> values) {
+            addCriterion("num not in", values, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverLike(String value) {
-            addCriterion("receiver like", value, "receiver");
+        public Criteria andNumBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("num between", value1, value2, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverNotLike(String value) {
-            addCriterion("receiver not like", value, "receiver");
+        public Criteria andNumNotBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("num not between", value1, value2, "num");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverIn(List<String> values) {
-            addCriterion("receiver in", values, "receiver");
+        public Criteria andUnitNumIsNull() {
+            addCriterion("unitNum is null");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverNotIn(List<String> values) {
-            addCriterion("receiver not in", values, "receiver");
+        public Criteria andUnitNumIsNotNull() {
+            addCriterion("unitNum is not null");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverBetween(String value1, String value2) {
-            addCriterion("receiver between", value1, value2, "receiver");
+        public Criteria andUnitNumEqualTo(BigDecimal value) {
+            addCriterion("unitNum =", value, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andReceiverNotBetween(String value1, String value2) {
-            addCriterion("receiver not between", value1, value2, "receiver");
+        public Criteria andUnitNumNotEqualTo(BigDecimal value) {
+            addCriterion("unitNum <>", value, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderIsNull() {
-            addCriterion("sender is null");
+        public Criteria andUnitNumGreaterThan(BigDecimal value) {
+            addCriterion("unitNum >", value, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderIsNotNull() {
-            addCriterion("sender is not null");
+        public Criteria andUnitNumGreaterThanOrEqualTo(BigDecimal value) {
+            addCriterion("unitNum >=", value, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderEqualTo(String value) {
-            addCriterion("sender =", value, "sender");
+        public Criteria andUnitNumLessThan(BigDecimal value) {
+            addCriterion("unitNum <", value, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderNotEqualTo(String value) {
-            addCriterion("sender <>", value, "sender");
+        public Criteria andUnitNumLessThanOrEqualTo(BigDecimal value) {
+            addCriterion("unitNum <=", value, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderGreaterThan(String value) {
-            addCriterion("sender >", value, "sender");
+        public Criteria andUnitNumIn(List<BigDecimal> values) {
+            addCriterion("unitNum in", values, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderGreaterThanOrEqualTo(String value) {
-            addCriterion("sender >=", value, "sender");
+        public Criteria andUnitNumNotIn(List<BigDecimal> values) {
+            addCriterion("unitNum not in", values, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderLessThan(String value) {
-            addCriterion("sender <", value, "sender");
+        public Criteria andUnitNumBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("unitNum between", value1, value2, "unitNum");
             return (Criteria) this;
         }
 
-        public Criteria andSenderLessThanOrEqualTo(String value) {
-            addCriterion("sender <=", value, "sender");
-            return (Criteria) this;
-        }
-
-        public Criteria andSenderLike(String value) {
-            addCriterion("sender like", value, "sender");
-            return (Criteria) this;
-        }
-
-        public Criteria andSenderNotLike(String value) {
-            addCriterion("sender not like", value, "sender");
-            return (Criteria) this;
-        }
-
-        public Criteria andSenderIn(List<String> values) {
-            addCriterion("sender in", values, "sender");
-            return (Criteria) this;
-        }
-
-        public Criteria andSenderNotIn(List<String> values) {
-            addCriterion("sender not in", values, "sender");
-            return (Criteria) this;
-        }
-
-        public Criteria andSenderBetween(String value1, String value2) {
-            addCriterion("sender between", value1, value2, "sender");
-            return (Criteria) this;
-        }
-
-        public Criteria andSenderNotBetween(String value1, String value2) {
-            addCriterion("sender not between", value1, value2, "sender");
+        public Criteria andUnitNumNotBetween(BigDecimal value1, BigDecimal value2) {
+            addCriterion("unitNum not between", value1, value2, "unitNum");
             return (Criteria) this;
         }
 
@@ -1365,6 +1226,11 @@ public class PmoutBillDetailExample {
             return (Criteria) this;
         }
 
+        public Criteria andProductNoLikeInsensitive(String value) {
+            addCriterion("upper(productNo) like", value.toUpperCase(), "productNo");
+            return (Criteria) this;
+        }
+
         public Criteria andPmNoLikeInsensitive(String value) {
             addCriterion("upper(pmNo) like", value.toUpperCase(), "pmNo");
             return (Criteria) this;
@@ -1375,8 +1241,8 @@ public class PmoutBillDetailExample {
             return (Criteria) this;
         }
 
-        public Criteria andPmBatchNoLikeInsensitive(String value) {
-            addCriterion("upper(pmBatchNo) like", value.toUpperCase(), "pmBatchNo");
+        public Criteria andTypeLikeInsensitive(String value) {
+            addCriterion("upper(type) like", value.toUpperCase(), "type");
             return (Criteria) this;
         }
 
@@ -1390,21 +1256,6 @@ public class PmoutBillDetailExample {
             return (Criteria) this;
         }
 
-        public Criteria andOutInfoLikeInsensitive(String value) {
-            addCriterion("upper(outInfo) like", value.toUpperCase(), "outInfo");
-            return (Criteria) this;
-        }
-
-        public Criteria andReceiverLikeInsensitive(String value) {
-            addCriterion("upper(receiver) like", value.toUpperCase(), "receiver");
-            return (Criteria) this;
-        }
-
-        public Criteria andSenderLikeInsensitive(String value) {
-            addCriterion("upper(sender) like", value.toUpperCase(), "sender");
-            return (Criteria) this;
-        }
-
         public Criteria andOperatorLikeInsensitive(String value) {
             addCriterion("upper(operator) like", value.toUpperCase(), "operator");
             return (Criteria) this;
@@ -1413,7 +1264,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_pmout_bill_detail
+     * This class corresponds to the database table t_productpm
      *
      * @mbg.generated do_not_delete_during_merge
      */
@@ -1426,7 +1277,7 @@ public class PmoutBillDetailExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_pmout_bill_detail
+     * This class corresponds to the database table t_productpm
      *
      * @mbg.generated
      */
