+++ src/main/resources/generatorConfig.xml
@@ -24,18 +24,18 @@
 		</javaTypeResolver>
 
 
-		<javaModelGenerator targetPackage="com.btjf.model.order" targetProject="src/main/java">
+		<javaModelGenerator targetPackage="com.btjf.model.emp" targetProject="src/main/java">
 			<property name="enableSubPackages" value="true" />
 			<property name="trimStrings" value="true" />
 		</javaModelGenerator>
-		<sqlMapGenerator targetPackage="com.btjf.mapper.order" targetProject="src/main/java">
+		<sqlMapGenerator targetPackage="com.btjf.mapper.emp" targetProject="src/main/java">
 			<property name="enableSubPackages" value="true" />
 		</sqlMapGenerator>
-		<javaClientGenerator targetPackage="com.btjf.mapper.order" targetProject="src/main/java" type="XMLMAPPER">
+		<javaClientGenerator targetPackage="com.btjf.mapper.emp" targetProject="src/main/java" type="XMLMAPPER">
 			<property name="enableSubPackages" value="true" />
 		</javaClientGenerator>
 
-		<table tableName="t_order" domainObjectName="Order">
+		<table tableName="t_emp_work" domainObjectName="EmpWork">
 		<property name="useActualColumnNames" value="true"/>
 		<generatedKey column="id" sqlStatement="MySql" identity="true"/>
 		<columnRenamingRule searchString="^F" replaceString="" />
